Revert "All `AuthProviders` must implement a non silent `check`." (#938)
Reverts databricks/databricks-vscode#936
