+++ .github/workflows/nightly-release.yml
@@ -2,7 +2,7 @@ name: Publish nightly release
 
 on:
     push:
-        branches: [main, bundle-integ]
+        branches: [main]
     workflow_dispatch:
 
 jobs:
@@ -24,8 +24,8 @@ jobs:
             - name: Update nightly release
               uses: softprops/action-gh-release@v1
               with:
-                  name: Nightly - ${{ github.ref_name }}
+                  name: Nightly
                   prerelease: true
-                  tag_name: nightly-${{ github.ref_name }}
+                  tag_name: nightly
                   token: ${{ secrets.GITHUB_TOKEN }}
                   files: "packages/databricks-vscode/databricks*/*.vsix"
