--- packages/databricks-vscode/.eslintrc.json
@@ -1,6 +1,0 @@
-{
-    "plugins": ["eslint-plugin-local-rules"],
-    "rules": {
-        "local-rules/mutex-synchronised-decorator": "error"
-    }
-}
