--- packages/databricks-vscode/eslint-local-rules/index.js
@@ -1,12 +1,0 @@
-/* eslint-disable */
-require("ts-node").register({
-    transpileOnly: true,
-    compilerOptions: {
-        module: "commonjs",
-    },
-});
-
-module.exports = {
-    "mutex-synchronised-decorator":
-        require("./rules/mutexSynchronisedDecorator").default,
-};
