+++ packages/databricks-vscode/package.json
@@ -666,7 +666,6 @@
         "package:cli:link": "rm -f ./bin/databricks && mkdir -p bin && ln -s ../../../../cli/cli bin/databricks",
         "package:wrappers:write": "ts-node ./scripts/writeIpynbWrapper.ts -s ./resources/python/notebook.workflow-wrapper.py -o ./resources/python/generated/notebook.workflow-wrapper.json",
         "package:jupyter-init-script:write": "ts-node ./scripts/writeJupyterInitFileWithVersion.ts",
-        "package:bundle-schema:write": "yarn package:cli:fetch && ts-node ./scripts/writeBundleSchema.ts ./bin/databricks ./src/bundle/BundleSchema.d.ts",
         "package:compile": "yarn run esbuild:base",
         "package:copy-webview-toolkit": "cp ./node_modules/@vscode/webview-ui-toolkit/dist/toolkit.js ./out/toolkit.js",
         "esbuild:base": "esbuild ./src/extension.ts --bundle --outfile=out/extension.js --external:vscode --format=cjs --platform=node --sourcemap --target=es2019",
@@ -688,7 +687,6 @@
     "dependencies": {
         "@databricks/databricks-sdk": "file:../../vendor/databricks-sdk.tgz",
         "@databricks/databricks-vscode-types": "workspace:^",
-        "@types/lodash": "^4.14.199",
         "@vscode/debugadapter": "^1.61.0",
         "@vscode/extension-telemetry": "^0.9.0",
         "@vscode/webview-ui-toolkit": "^1.2.2",
@@ -704,9 +702,8 @@
         "@sinonjs/fake-timers": "^11.2.2",
         "@types/bcryptjs": "^2.4.2",
         "@types/chai": "^4.3.5",
-        "@types/eslint": "^8.44.6",
         "@types/fs-extra": "^11.0.1",
-        "@types/mocha": "^10.0.2",
+        "@types/mocha": "^10.0.1",
         "@types/mock-require": "^2.0.1",
         "@types/node": "^20.4.2",
         "@types/sinonjs__fake-timers": "^8.1.2",
@@ -716,7 +713,6 @@
         "@types/yargs": "^17.0.24",
         "@typescript-eslint/eslint-plugin": "^6.0.0",
         "@typescript-eslint/parser": "^6.0.0",
-        "@typescript-eslint/utils": "^6.9.0",
         "@vscode/test-electron": "^2.3.3",
         "@wdio/cli": "^8.12.2",
         "@wdio/local-runner": "^8.12.1",
@@ -726,10 +722,8 @@
         "chai": "^4.3.7",
         "esbuild": "^0.19.4",
         "eslint": "^8.51.0",
-        "eslint-plugin-local-rules": "^2.0.0",
         "fs-extra": "^11.1.1",
-        "glob": "^10.3.10",
-        "json-schema-to-typescript": "^13.1.1",
+        "glob": "^10.3.3",
         "mocha": "^10.2.0",
         "mock-require": "^3.0.3",
         "nyc": "^15.1.0",
