--- packages/databricks-vscode/src/bundle/BundleSchema.d.ts
@@ -1,5541 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-/**
- * Root of the bundle config
- */
-export interface BundleSchema {
-    /**
-     * A description of all code artifacts in this bundle.
-     */
-    artifacts?: {
-        [k: string]: {
-            build: string;
-            files: {
-                source: string;
-            }[];
-            path: string;
-            type: string;
-        };
-    };
-    /**
-     * The details for this bundle.
-     */
-    bundle?: {
-        compute_id?: string;
-        git?: {
-            branch?: string;
-            origin_url?: string;
-        };
-        /**
-         * The name of the bundle.
-         */
-        name: string;
-    };
-    environments?: {
-        [k: string]: {
-            artifacts?: {
-                [k: string]: {
-                    build: string;
-                    files: {
-                        source: string;
-                    }[];
-                    path: string;
-                    type: string;
-                };
-            };
-            bundle?: {
-                compute_id?: string;
-                git?: {
-                    branch?: string;
-                    origin_url?: string;
-                };
-                name: string;
-            };
-            compute_id?: string;
-            default?: boolean;
-            git?: {
-                branch?: string;
-                origin_url?: string;
-            };
-            mode?: string;
-            resources?: {
-                experiments?: {
-                    [k: string]: {
-                        artifact_location?: string;
-                        creation_time?: number;
-                        experiment_id?: string;
-                        last_update_time?: number;
-                        lifecycle_stage?: string;
-                        name?: string;
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                        tags?: {
-                            key?: string;
-                            value?: string;
-                        }[];
-                    };
-                };
-                jobs?: {
-                    [k: string]: {
-                        compute?: {
-                            compute_key: string;
-                            spec: {
-                                kind?: string;
-                            };
-                        }[];
-                        continuous?: {
-                            pause_status?: string;
-                        };
-                        email_notifications?: {
-                            no_alert_for_skipped_runs?: boolean;
-                            on_duration_warning_threshold_exceeded?: string[];
-                            on_failure?: string[];
-                            on_start?: string[];
-                            on_success?: string[];
-                        };
-                        format?: string;
-                        git_source?: {
-                            git_branch?: string;
-                            git_commit?: string;
-                            git_provider: string;
-                            git_snapshot?: {
-                                used_commit?: string;
-                            };
-                            git_tag?: string;
-                            git_url: string;
-                            job_source?: {
-                                dirty_state?: string;
-                                import_from_git_branch: string;
-                                job_config_path: string;
-                            };
-                        };
-                        health?: {
-                            rules?: {
-                                metric?: string;
-                                op?: string;
-                                value?: number;
-                            }[];
-                        };
-                        job_clusters?: {
-                            job_cluster_key: string;
-                            new_cluster?: {
-                                autoscale?: {
-                                    max_workers: number;
-                                    min_workers: number;
-                                };
-                                autotermination_minutes?: number;
-                                aws_attributes?: {
-                                    availability?: string;
-                                    ebs_volume_count?: number;
-                                    ebs_volume_iops?: number;
-                                    ebs_volume_size?: number;
-                                    ebs_volume_throughput?: number;
-                                    ebs_volume_type?: string;
-                                    first_on_demand?: number;
-                                    instance_profile_arn?: string;
-                                    spot_bid_price_percent?: number;
-                                    zone_id?: string;
-                                };
-                                azure_attributes?: {
-                                    availability?: string;
-                                    first_on_demand?: number;
-                                    log_analytics_info?: {
-                                        log_analytics_primary_key?: string;
-                                        log_analytics_workspace_id?: string;
-                                    };
-                                    spot_bid_max_price?: number;
-                                };
-                                cluster_log_conf?: {
-                                    dbfs?: {
-                                        destination?: string;
-                                    };
-                                    s3?: {
-                                        canned_acl?: string;
-                                        destination?: string;
-                                        enable_encryption?: boolean;
-                                        encryption_type?: string;
-                                        endpoint?: string;
-                                        kms_key?: string;
-                                        region?: string;
-                                    };
-                                };
-                                cluster_name?: string;
-                                cluster_source?: string;
-                                custom_tags?: {
-                                    [k: string]: string;
-                                };
-                                data_security_mode?: string;
-                                docker_image?: {
-                                    basic_auth?: {
-                                        password?: string;
-                                        username?: string;
-                                    };
-                                    url?: string;
-                                };
-                                driver_instance_pool_id?: string;
-                                driver_node_type_id?: string;
-                                enable_elastic_disk?: boolean;
-                                enable_local_disk_encryption?: boolean;
-                                gcp_attributes?: {
-                                    availability?: string;
-                                    boot_disk_size?: number;
-                                    google_service_account?: string;
-                                    local_ssd_count?: number;
-                                };
-                                init_scripts?: {
-                                    dbfs?: {
-                                        destination?: string;
-                                    };
-                                    s3?: {
-                                        canned_acl?: string;
-                                        destination?: string;
-                                        enable_encryption?: boolean;
-                                        encryption_type?: string;
-                                        endpoint?: string;
-                                        kms_key?: string;
-                                        region?: string;
-                                    };
-                                    volumes?: {
-                                        destination?: string;
-                                    };
-                                    workspace?: {
-                                        destination?: string;
-                                    };
-                                }[];
-                                instance_pool_id?: string;
-                                node_type_id?: string;
-                                num_workers?: number;
-                                policy_id?: string;
-                                runtime_engine?: string;
-                                single_user_name?: string;
-                                spark_conf?: {
-                                    [k: string]: string;
-                                };
-                                spark_env_vars?: {
-                                    [k: string]: string;
-                                };
-                                spark_version?: string;
-                                ssh_public_keys?: string[];
-                                workload_type?: {
-                                    clients?: {
-                                        jobs?: boolean;
-                                        notebooks?: boolean;
-                                    };
-                                };
-                            };
-                        }[];
-                        max_concurrent_runs?: number;
-                        name?: string;
-                        notification_settings?: {
-                            no_alert_for_canceled_runs?: boolean;
-                            no_alert_for_skipped_runs?: boolean;
-                        };
-                        parameters?: {
-                            default: string;
-                            name: string;
-                        }[];
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                        run_as?: {
-                            service_principal_name?: string;
-                            user_name?: string;
-                        };
-                        schedule?: {
-                            pause_status?: string;
-                            quartz_cron_expression: string;
-                            timezone_id: string;
-                        };
-                        tags?: {
-                            [k: string]: string;
-                        };
-                        tasks?: {
-                            compute_key?: string;
-                            condition_task?: {
-                                left?: string;
-                                op?: string;
-                                right?: string;
-                            };
-                            dbt_task?: {
-                                catalog?: string;
-                                commands: string[];
-                                profiles_directory?: string;
-                                project_directory?: string;
-                                schema?: string;
-                                warehouse_id?: string;
-                            };
-                            depends_on?: {
-                                outcome?: string;
-                                task_key: string;
-                            }[];
-                            description?: string;
-                            email_notifications?: {
-                                on_duration_warning_threshold_exceeded?: string[];
-                                on_failure?: string[];
-                                on_start?: string[];
-                                on_success?: string[];
-                            };
-                            existing_cluster_id?: string;
-                            health?: {
-                                rules?: {
-                                    metric?: string;
-                                    op?: string;
-                                    value?: number;
-                                }[];
-                            };
-                            job_cluster_key?: string;
-                            libraries?: {
-                                cran?: {
-                                    package: string;
-                                    repo?: string;
-                                };
-                                egg?: string;
-                                jar?: string;
-                                maven?: {
-                                    coordinates: string;
-                                    exclusions?: string[];
-                                    repo?: string;
-                                };
-                                pypi?: {
-                                    package: string;
-                                    repo?: string;
-                                };
-                                whl?: string;
-                            }[];
-                            max_retries?: number;
-                            min_retry_interval_millis?: number;
-                            new_cluster?: {
-                                autoscale?: {
-                                    max_workers: number;
-                                    min_workers: number;
-                                };
-                                autotermination_minutes?: number;
-                                aws_attributes?: {
-                                    availability?: string;
-                                    ebs_volume_count?: number;
-                                    ebs_volume_iops?: number;
-                                    ebs_volume_size?: number;
-                                    ebs_volume_throughput?: number;
-                                    ebs_volume_type?: string;
-                                    first_on_demand?: number;
-                                    instance_profile_arn?: string;
-                                    spot_bid_price_percent?: number;
-                                    zone_id?: string;
-                                };
-                                azure_attributes?: {
-                                    availability?: string;
-                                    first_on_demand?: number;
-                                    log_analytics_info?: {
-                                        log_analytics_primary_key?: string;
-                                        log_analytics_workspace_id?: string;
-                                    };
-                                    spot_bid_max_price?: number;
-                                };
-                                cluster_log_conf?: {
-                                    dbfs?: {
-                                        destination?: string;
-                                    };
-                                    s3?: {
-                                        canned_acl?: string;
-                                        destination?: string;
-                                        enable_encryption?: boolean;
-                                        encryption_type?: string;
-                                        endpoint?: string;
-                                        kms_key?: string;
-                                        region?: string;
-                                    };
-                                };
-                                cluster_name?: string;
-                                cluster_source?: string;
-                                custom_tags?: {
-                                    [k: string]: string;
-                                };
-                                data_security_mode?: string;
-                                docker_image?: {
-                                    basic_auth?: {
-                                        password?: string;
-                                        username?: string;
-                                    };
-                                    url?: string;
-                                };
-                                driver_instance_pool_id?: string;
-                                driver_node_type_id?: string;
-                                enable_elastic_disk?: boolean;
-                                enable_local_disk_encryption?: boolean;
-                                gcp_attributes?: {
-                                    availability?: string;
-                                    boot_disk_size?: number;
-                                    google_service_account?: string;
-                                    local_ssd_count?: number;
-                                };
-                                init_scripts?: {
-                                    dbfs?: {
-                                        destination?: string;
-                                    };
-                                    s3?: {
-                                        canned_acl?: string;
-                                        destination?: string;
-                                        enable_encryption?: boolean;
-                                        encryption_type?: string;
-                                        endpoint?: string;
-                                        kms_key?: string;
-                                        region?: string;
-                                    };
-                                    volumes?: {
-                                        destination?: string;
-                                    };
-                                    workspace?: {
-                                        destination?: string;
-                                    };
-                                }[];
-                                instance_pool_id?: string;
-                                node_type_id?: string;
-                                num_workers?: number;
-                                policy_id?: string;
-                                runtime_engine?: string;
-                                single_user_name?: string;
-                                spark_conf?: {
-                                    [k: string]: string;
-                                };
-                                spark_env_vars?: {
-                                    [k: string]: string;
-                                };
-                                spark_version?: string;
-                                ssh_public_keys?: string[];
-                                workload_type?: {
-                                    clients?: {
-                                        jobs?: boolean;
-                                        notebooks?: boolean;
-                                    };
-                                };
-                            };
-                            notebook_task?: {
-                                base_parameters?: {
-                                    [k: string]: string;
-                                };
-                                notebook_path: string;
-                                source?: string;
-                            };
-                            notification_settings?: {
-                                alert_on_last_attempt?: boolean;
-                                no_alert_for_canceled_runs?: boolean;
-                                no_alert_for_skipped_runs?: boolean;
-                            };
-                            pipeline_task?: {
-                                full_refresh?: boolean;
-                                pipeline_id?: string;
-                            };
-                            python_wheel_task?: {
-                                entry_point?: string;
-                                named_parameters?: {
-                                    [k: string]: string;
-                                };
-                                package_name?: string;
-                                parameters?: string[];
-                            };
-                            retry_on_timeout?: boolean;
-                            run_if?: string;
-                            run_job_task?: {
-                                job_id: number;
-                                job_parameters?: unknown;
-                            };
-                            spark_jar_task?: {
-                                jar_uri?: string;
-                                main_class_name?: string;
-                                parameters?: string[];
-                            };
-                            spark_python_task?: {
-                                parameters?: string[];
-                                python_file: string;
-                                source?: string;
-                            };
-                            spark_submit_task?: {
-                                parameters?: string[];
-                            };
-                            sql_task?: {
-                                alert?: {
-                                    alert_id: string;
-                                    pause_subscriptions?: boolean;
-                                    subscriptions?: {
-                                        destination_id?: string;
-                                        user_name?: string;
-                                    }[];
-                                };
-                                dashboard?: {
-                                    custom_subject?: string;
-                                    dashboard_id: string;
-                                    pause_subscriptions?: boolean;
-                                    subscriptions?: {
-                                        destination_id?: string;
-                                        user_name?: string;
-                                    }[];
-                                };
-                                file?: {
-                                    path: string;
-                                };
-                                parameters?: {
-                                    [k: string]: string;
-                                };
-                                query?: {
-                                    query_id: string;
-                                };
-                                warehouse_id: string;
-                            };
-                            task_key: string;
-                            timeout_seconds?: number;
-                        }[];
-                        timeout_seconds?: number;
-                        trigger?: {
-                            file_arrival?: {
-                                min_time_between_triggers_seconds?: number;
-                                url?: string;
-                                wait_after_last_change_seconds?: number;
-                            };
-                            pause_status?: string;
-                        };
-                        webhook_notifications?: {
-                            on_duration_warning_threshold_exceeded?: {
-                                id?: string;
-                            }[];
-                            on_failure?: {
-                                id?: string;
-                            }[];
-                            on_start?: {
-                                id?: string;
-                            }[];
-                            on_success?: {
-                                id?: string;
-                            }[];
-                        };
-                    };
-                };
-                model_serving_endpoints?: {
-                    [k: string]: {
-                        config: {
-                            served_models: {
-                                environment_vars?: {
-                                    [k: string]: string;
-                                };
-                                instance_profile_arn?: string;
-                                model_name: string;
-                                model_version: string;
-                                name?: string;
-                                scale_to_zero_enabled: boolean;
-                                workload_size: string;
-                            }[];
-                            traffic_config?: {
-                                routes?: {
-                                    served_model_name: string;
-                                    traffic_percentage: number;
-                                }[];
-                            };
-                        };
-                        name: string;
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                    };
-                };
-                models?: {
-                    [k: string]: {
-                        creation_timestamp?: number;
-                        description?: string;
-                        last_updated_timestamp?: number;
-                        latest_versions?: {
-                            creation_timestamp?: number;
-                            current_stage?: string;
-                            description?: string;
-                            last_updated_timestamp?: number;
-                            name?: string;
-                            run_id?: string;
-                            run_link?: string;
-                            source?: string;
-                            status?: string;
-                            status_message?: string;
-                            tags?: {
-                                key?: string;
-                                value?: string;
-                            }[];
-                            user_id?: string;
-                            version?: string;
-                        }[];
-                        name?: string;
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                        tags?: {
-                            key?: string;
-                            value?: string;
-                        }[];
-                        user_id?: string;
-                    };
-                };
-                pipelines?: {
-                    [k: string]: {
-                        catalog?: string;
-                        channel?: string;
-                        clusters?: {
-                            apply_policy_default_values?: boolean;
-                            autoscale?: {
-                                max_workers: number;
-                                min_workers: number;
-                            };
-                            aws_attributes?: {
-                                availability?: string;
-                                ebs_volume_count?: number;
-                                ebs_volume_iops?: number;
-                                ebs_volume_size?: number;
-                                ebs_volume_throughput?: number;
-                                ebs_volume_type?: string;
-                                first_on_demand?: number;
-                                instance_profile_arn?: string;
-                                spot_bid_price_percent?: number;
-                                zone_id?: string;
-                            };
-                            azure_attributes?: {
-                                availability?: string;
-                                first_on_demand?: number;
-                                log_analytics_info?: {
-                                    log_analytics_primary_key?: string;
-                                    log_analytics_workspace_id?: string;
-                                };
-                                spot_bid_max_price?: number;
-                            };
-                            cluster_log_conf?: {
-                                dbfs?: {
-                                    destination?: string;
-                                };
-                                s3?: {
-                                    canned_acl?: string;
-                                    destination?: string;
-                                    enable_encryption?: boolean;
-                                    encryption_type?: string;
-                                    endpoint?: string;
-                                    kms_key?: string;
-                                    region?: string;
-                                };
-                            };
-                            custom_tags?: {
-                                [k: string]: string;
-                            };
-                            driver_instance_pool_id?: string;
-                            driver_node_type_id?: string;
-                            gcp_attributes?: {
-                                availability?: string;
-                                boot_disk_size?: number;
-                                google_service_account?: string;
-                                local_ssd_count?: number;
-                            };
-                            instance_pool_id?: string;
-                            label?: string;
-                            node_type_id?: string;
-                            num_workers?: number;
-                            policy_id?: string;
-                            spark_conf?: {
-                                [k: string]: string;
-                            };
-                            spark_env_vars?: {
-                                [k: string]: string;
-                            };
-                            ssh_public_keys?: string[];
-                        }[];
-                        configuration?: {
-                            [k: string]: string;
-                        };
-                        continuous?: boolean;
-                        development?: boolean;
-                        edition?: string;
-                        filters?: {
-                            exclude?: string[];
-                            include?: string[];
-                        };
-                        id?: string;
-                        libraries?: {
-                            file?: {
-                                path?: string;
-                            };
-                            jar?: string;
-                            maven?: {
-                                coordinates: string;
-                                exclusions?: string[];
-                                repo?: string;
-                            };
-                            notebook?: {
-                                path?: string;
-                            };
-                        }[];
-                        name?: string;
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                        photon?: boolean;
-                        serverless?: boolean;
-                        storage?: string;
-                        target?: string;
-                        trigger?: {
-                            cron?: {
-                                quartz_cron_schedule?: string;
-                                timezone_id?: string;
-                            };
-                            manual?: unknown;
-                        };
-                    };
-                };
-            };
-            run_as?: {
-                service_principal_name?: string;
-                user_name?: string;
-            };
-            variables?: {
-                [k: string]: string;
-            };
-            workspace?: {
-                artifact_path?: string;
-                auth_type?: string;
-                azure_client_id?: string;
-                azure_environment?: string;
-                azure_login_app_id?: string;
-                azure_tenant_id?: string;
-                azure_use_msi?: boolean;
-                azure_workspace_resource_id?: string;
-                client_id?: string;
-                file_path?: string;
-                google_service_account?: string;
-                host?: string;
-                profile?: string;
-                root_path?: string;
-                state_path?: string;
-            };
-        };
-    };
-    experimental?: {
-        python_wheel_wrapper?: boolean;
-        scripts?: {
-            [k: string]: string;
-        };
-    };
-    /**
-     * A list of glob patterns of files to load and merge into the this configuration. Defaults to no files being included.
-     */
-    include?: string[];
-    /**
-     * Collection of Databricks resources to deploy.
-     */
-    resources?: {
-        /**
-         * List of MLflow experiments
-         */
-        experiments?: {
-            [k: string]: {
-                /**
-                 * Location where artifacts for the experiment are stored.
-                 */
-                artifact_location?: string;
-                /**
-                 * Creation time
-                 */
-                creation_time?: number;
-                /**
-                 * Unique identifier for the experiment.
-                 */
-                experiment_id?: string;
-                /**
-                 * Last update time
-                 */
-                last_update_time?: number;
-                /**
-                 * Current life cycle stage of the experiment: "active" or "deleted".
-                 * Deleted experiments are not returned by APIs.
-                 */
-                lifecycle_stage?: string;
-                /**
-                 * Human readable name that identifies the experiment.
-                 */
-                name?: string;
-                permissions?: {
-                    group_name?: string;
-                    level: string;
-                    service_principal_name?: string;
-                    user_name?: string;
-                }[];
-                /**
-                 * Tags: Additional metadata key-value pairs.
-                 */
-                tags?: {
-                    /**
-                     * The tag key.
-                     */
-                    key?: string;
-                    /**
-                     * The tag value.
-                     */
-                    value?: string;
-                }[];
-            };
-        };
-        /**
-         * List of Databricks jobs
-         */
-        jobs?: {
-            [k: string]: {
-                /**
-                 * A list of compute requirements that can be referenced by tasks of this job.
-                 */
-                compute?: {
-                    /**
-                     * A unique name for the compute requirement. This field is required and must be unique within the job.
-                     * `JobTaskSettings` may refer to this field to determine the compute requirements for the task execution.
-                     */
-                    compute_key: string;
-                    spec: {
-                        /**
-                         * The kind of compute described by this compute specification.
-                         */
-                        kind?: string;
-                    };
-                }[];
-                /**
-                 * An optional continuous property for this job. The continuous property will ensure that there is always one run executing. Only one of `schedule` and `continuous` can be used.
-                 */
-                continuous?: {
-                    /**
-                     * Whether this trigger is paused or not.
-                     */
-                    pause_status?: string;
-                };
-                /**
-                 * An optional set of email addresses that is notified when runs of this job begin or complete as well as when this job is deleted. The default behavior is to not send any emails.
-                 */
-                email_notifications?: {
-                    /**
-                     * If true, do not send email to recipients specified in `on_failure` if the run is skipped.
-                     */
-                    no_alert_for_skipped_runs?: boolean;
-                    on_duration_warning_threshold_exceeded?: string[];
-                    /**
-                     * A list of email addresses to be notified when a run unsuccessfully completes. A run is considered to have completed unsuccessfully if it ends with an `INTERNAL_ERROR` `life_cycle_state` or a `FAILED`, or `TIMED_OUT` result_state. If this is not specified on job creation, reset, or update the list is empty, and notifications are not sent.
-                     */
-                    on_failure?: string[];
-                    /**
-                     * A list of email addresses to be notified when a run begins. If not specified on job creation, reset, or update, the list is empty, and notifications are not sent.
-                     */
-                    on_start?: string[];
-                    /**
-                     * A list of email addresses to be notified when a run successfully completes. A run is considered to have completed successfully if it ends with a `TERMINATED` `life_cycle_state` and a `SUCCESS` result_state. If not specified on job creation, reset, or update, the list is empty, and notifications are not sent.
-                     */
-                    on_success?: string[];
-                };
-                /**
-                 * Used to tell what is the format of the job. This field is ignored in Create/Update/Reset calls. When using the Jobs API 2.1 this value is always set to `"MULTI_TASK"`.
-                 */
-                format?: string;
-                /**
-                 * An optional specification for a remote repository containing the notebooks used by this job's notebook tasks.
-                 */
-                git_source?: {
-                    /**
-                     * Name of the branch to be checked out and used by this job.
-                     * This field cannot be specified in conjunction with git_tag or git_commit.
-                     *
-                     * The maximum length is 255 characters.
-                     *
-                     */
-                    git_branch?: string;
-                    /**
-                     * Commit to be checked out and used by this job. This field cannot be specified in conjunction with git_branch or git_tag.
-                     * The maximum length is 64 characters.
-                     */
-                    git_commit?: string;
-                    /**
-                     * Unique identifier of the service used to host the Git repository. The value is case insensitive.
-                     */
-                    git_provider: string;
-                    git_snapshot?: {
-                        /**
-                         * Commit that was used to execute the run. If git_branch was specified, this points to the HEAD of the branch at the time of the run; if git_tag was specified, this points to the commit the tag points to.
-                         */
-                        used_commit?: string;
-                    };
-                    /**
-                     * Name of the tag to be checked out and used by this job.
-                     * This field cannot be specified in conjunction with git_branch or git_commit.
-                     *
-                     * The maximum length is 255 characters.
-                     *
-                     */
-                    git_tag?: string;
-                    /**
-                     * URL of the repository to be cloned by this job.
-                     * The maximum length is 300 characters.
-                     */
-                    git_url: string;
-                    job_source?: {
-                        dirty_state?: string;
-                        import_from_git_branch: string;
-                        job_config_path: string;
-                    };
-                };
-                health?: {
-                    rules?: {
-                        metric?: string;
-                        op?: string;
-                        value?: number;
-                    }[];
-                };
-                /**
-                 * A list of job cluster specifications that can be shared and reused by tasks of this job. Libraries cannot be declared in a shared job cluster. You must declare dependent libraries in task settings.
-                 */
-                job_clusters?: {
-                    /**
-                     * A unique name for the job cluster. This field is required and must be unique within the job.
-                     * `JobTaskSettings` may refer to this field to determine which cluster to launch for the task execution.
-                     */
-                    job_cluster_key: string;
-                    /**
-                     * If new_cluster, a description of a cluster that is created for only for this task.
-                     */
-                    new_cluster?: {
-                        /**
-                         * Parameters needed in order to automatically scale clusters up and down based on load.
-                         * Note: autoscaling works best with DB runtime versions 3.0 or later.
-                         */
-                        autoscale?: {
-                            /**
-                             * The maximum number of workers to which the cluster can scale up when overloaded.
-                             * Note that `max_workers` must be strictly greater than `min_workers`.
-                             */
-                            max_workers: number;
-                            /**
-                             * The minimum number of workers to which the cluster can scale down when underutilized.
-                             * It is also the initial number of workers the cluster will have after creation.
-                             */
-                            min_workers: number;
-                        };
-                        /**
-                         * Automatically terminates the cluster after it is inactive for this time in minutes. If not set,
-                         * this cluster will not be automatically terminated. If specified, the threshold must be between
-                         * 10 and 10000 minutes.
-                         * Users can also set this value to 0 to explicitly disable automatic termination.
-                         */
-                        autotermination_minutes?: number;
-                        /**
-                         * Attributes related to clusters running on Amazon Web Services.
-                         * If not specified at cluster creation, a set of default values will be used.
-                         */
-                        aws_attributes?: {
-                            availability?: string;
-                            /**
-                             * The number of volumes launched for each instance. Users can choose up to 10 volumes.
-                             * This feature is only enabled for supported node types. Legacy node types cannot specify
-                             * custom EBS volumes.
-                             * For node types with no instance store, at least one EBS volume needs to be specified;
-                             * otherwise, cluster creation will fail.
-                             *
-                             * These EBS volumes will be mounted at `/ebs0`, `/ebs1`, and etc.
-                             * Instance store volumes will be mounted at `/local_disk0`, `/local_disk1`, and etc.
-                             *
-                             * If EBS volumes are attached, Databricks will configure Spark to use only the EBS volumes for
-                             * scratch storage because heterogenously sized scratch devices can lead to inefficient disk
-                             * utilization. If no EBS volumes are attached, Databricks will configure Spark to use instance
-                             * store volumes.
-                             *
-                             * Please note that if EBS volumes are specified, then the Spark configuration `spark.local.dir`
-                             * will be overridden.
-                             */
-                            ebs_volume_count?: number;
-                            /**
-                             * <needs content added>
-                             */
-                            ebs_volume_iops?: number;
-                            /**
-                             * The size of each EBS volume (in GiB) launched for each instance. For general purpose
-                             * SSD, this value must be within the range 100 - 4096. For throughput optimized HDD,
-                             * this value must be within the range 500 - 4096.
-                             */
-                            ebs_volume_size?: number;
-                            /**
-                             * <needs content added>
-                             */
-                            ebs_volume_throughput?: number;
-                            ebs_volume_type?: string;
-                            /**
-                             * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                             * If this value is greater than 0, the cluster driver node in particular will be placed on an
-                             * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                             * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                             * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                             * be placed on `availability` instances. Note that this value does not affect
-                             * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                             */
-                            first_on_demand?: number;
-                            /**
-                             * Nodes for this cluster will only be placed on AWS instances with this instance profile. If
-                             * ommitted, nodes will be placed on instances without an IAM instance profile. The instance
-                             * profile must have previously been added to the Databricks environment by an account
-                             * administrator.
-                             *
-                             * This feature may only be available to certain customer plans.
-                             *
-                             * If this field is ommitted, we will pull in the default from the conf if it exists.
-                             */
-                            instance_profile_arn?: string;
-                            /**
-                             * The bid price for AWS spot instances, as a percentage of the corresponding instance type's
-                             * on-demand price.
-                             * For example, if this field is set to 50, and the cluster needs a new `r3.xlarge` spot
-                             * instance, then the bid price is half of the price of
-                             * on-demand `r3.xlarge` instances. Similarly, if this field is set to 200, the bid price is twice
-                             * the price of on-demand `r3.xlarge` instances. If not specified, the default value is 100.
-                             * When spot instances are requested for this cluster, only spot instances whose bid price
-                             * percentage matches this field will be considered.
-                             * Note that, for safety, we enforce this field to be no more than 10000.
-                             *
-                             * The default value and documentation here should be kept consistent with
-                             * CommonConf.defaultSpotBidPricePercent and CommonConf.maxSpotBidPricePercent.
-                             */
-                            spot_bid_price_percent?: number;
-                            /**
-                             * Identifier for the availability zone/datacenter in which the cluster resides.
-                             * This string will be of a form like "us-west-2a". The provided availability
-                             * zone must be in the same region as the Databricks deployment. For example, "us-west-2a"
-                             * is not a valid zone id if the Databricks deployment resides in the "us-east-1" region.
-                             * This is an optional field at cluster creation, and if not specified, a default zone will be used.
-                             * If the zone specified is "auto", will try to place cluster in a zone with high availability,
-                             * and will retry placement in a different AZ if there is not enough capacity.
-                             * See [[AutoAZHelper.scala]] for more details.
-                             * The list of available zones as well as the default value can be found by using the
-                             * `List Zones`_ method.
-                             */
-                            zone_id?: string;
-                        };
-                        /**
-                         * Attributes related to clusters running on Microsoft Azure.
-                         * If not specified at cluster creation, a set of default values will be used.
-                         */
-                        azure_attributes?: {
-                            availability?: string;
-                            /**
-                             * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                             * This value should be greater than 0, to make sure the cluster driver node is placed on an
-                             * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                             * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                             * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                             * be placed on `availability` instances. Note that this value does not affect
-                             * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                             */
-                            first_on_demand?: number;
-                            /**
-                             * Defines values necessary to configure and run Azure Log Analytics agent
-                             */
-                            log_analytics_info?: {
-                                /**
-                                 * <needs content added>
-                                 */
-                                log_analytics_primary_key?: string;
-                                /**
-                                 * <needs content added>
-                                 */
-                                log_analytics_workspace_id?: string;
-                            };
-                            /**
-                             * The max bid price to be used for Azure spot instances.
-                             * The Max price for the bid cannot be higher than the on-demand price of the instance.
-                             * If not specified, the default value is -1, which specifies that the instance cannot be evicted
-                             * on the basis of price, and only on the basis of availability. Further, the value should > 0 or -1.
-                             */
-                            spot_bid_max_price?: number;
-                        };
-                        /**
-                         * The configuration for delivering spark logs to a long-term storage destination.
-                         * Two kinds of destinations (dbfs and s3) are supported. Only one destination can be specified
-                         * for one cluster. If the conf is given, the logs will be delivered to the destination every
-                         * `5 mins`. The destination of driver logs is `$destination/$clusterId/driver`, while
-                         * the destination of executor logs is `$destination/$clusterId/executor`.
-                         */
-                        cluster_log_conf?: {
-                            /**
-                             * destination needs to be provided. e.g.
-                             * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                             */
-                            dbfs?: {
-                                /**
-                                 * dbfs destination, e.g. `dbfs:/my/path`
-                                 */
-                                destination?: string;
-                            };
-                            /**
-                             * destination and either the region or endpoint need to be provided. e.g.
-                             * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                             * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                             * `instance_profile_arn` has permission to write data to the s3 destination.
-                             */
-                            s3?: {
-                                /**
-                                 * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                                 * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                                 * the destination bucket and prefix. The full list of possible canned acl can be found at
-                                 * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                                 * Please also note that by default only the object owner gets full controls. If you are using cross account
-                                 * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                                 * read the logs.
-                                 */
-                                canned_acl?: string;
-                                /**
-                                 * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                                 * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                                 * destination. Please also note that you cannot use AWS keys to deliver logs.
-                                 */
-                                destination?: string;
-                                /**
-                                 * (Optional) Flag to enable server side encryption, `false` by default.
-                                 */
-                                enable_encryption?: boolean;
-                                /**
-                                 * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                                 * encryption is enabled and the default type is `sse-s3`.
-                                 */
-                                encryption_type?: string;
-                                /**
-                                 * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                                 * If both are set, endpoint will be used.
-                                 */
-                                endpoint?: string;
-                                /**
-                                 * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                                 */
-                                kms_key?: string;
-                                /**
-                                 * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                                 * endpoint will be used.
-                                 */
-                                region?: string;
-                            };
-                        };
-                        /**
-                         * Cluster name requested by the user. This doesn't have to be unique.
-                         * If not specified at creation, the cluster name will be an empty string.
-                         *
-                         */
-                        cluster_name?: string;
-                        cluster_source?: string;
-                        /**
-                         * Additional tags for cluster resources. Databricks will tag all cluster resources (e.g., AWS
-                         * instances and EBS volumes) with these tags in addition to `default_tags`. Notes:
-                         *
-                         * - Currently, Databricks allows at most 45 custom tags
-                         *
-                         * - Clusters can only reuse cloud resources if the resources' tags are a subset of the cluster tags
-                         */
-                        custom_tags?: {
-                            [k: string]: string;
-                        };
-                        data_security_mode?: string;
-                        docker_image?: {
-                            basic_auth?: {
-                                /**
-                                 * Password of the user
-                                 */
-                                password?: string;
-                                /**
-                                 * Name of the user
-                                 */
-                                username?: string;
-                            };
-                            /**
-                             * URL of the docker image.
-                             */
-                            url?: string;
-                        };
-                        /**
-                         * The optional ID of the instance pool for the driver of the cluster belongs.
-                         * The pool cluster uses the instance pool with id (instance_pool_id) if the driver pool is not
-                         * assigned.
-                         */
-                        driver_instance_pool_id?: string;
-                        /**
-                         * The node type of the Spark driver. Note that this field is optional;
-                         * if unset, the driver node type will be set as the same value
-                         * as `node_type_id` defined above.
-                         *
-                         */
-                        driver_node_type_id?: string;
-                        /**
-                         * Autoscaling Local Storage: when enabled, this cluster will dynamically acquire additional disk
-                         * space when its Spark workers are running low on disk space. This feature requires specific AWS
-                         * permissions to function correctly - refer to the User Guide for more details.
-                         */
-                        enable_elastic_disk?: boolean;
-                        /**
-                         * Whether to enable LUKS on cluster VMs' local disks
-                         */
-                        enable_local_disk_encryption?: boolean;
-                        /**
-                         * Attributes related to clusters running on Google Cloud Platform.
-                         * If not specified at cluster creation, a set of default values will be used.
-                         */
-                        gcp_attributes?: {
-                            availability?: string;
-                            /**
-                             * boot disk size in GB
-                             */
-                            boot_disk_size?: number;
-                            /**
-                             * If provided, the cluster will impersonate the google service account when accessing
-                             * gcloud services (like GCS). The google service account
-                             * must have previously been added to the Databricks environment by an account
-                             * administrator.
-                             */
-                            google_service_account?: string;
-                            /**
-                             * If provided, each node (workers and driver) in the cluster will have this number of local SSDs attached. Each local SSD is 375GB in size. Refer to [GCP documentation](https://cloud.google.com/compute/docs/disks/local-ssd#choose_number_local_ssds) for the supported number of local SSDs for each instance type.
-                             */
-                            local_ssd_count?: number;
-                        };
-                        /**
-                         * The configuration for storing init scripts. Any number of destinations can be specified. The scripts are executed sequentially in the order provided. If `cluster_log_conf` is specified, init script logs are sent to `<destination>/<cluster-ID>/init_scripts`.
-                         */
-                        init_scripts?: {
-                            /**
-                             * destination needs to be provided. e.g.
-                             * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                             */
-                            dbfs?: {
-                                /**
-                                 * dbfs destination, e.g. `dbfs:/my/path`
-                                 */
-                                destination?: string;
-                            };
-                            /**
-                             * destination and either the region or endpoint need to be provided. e.g.
-                             * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                             * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                             * `instance_profile_arn` has permission to write data to the s3 destination.
-                             */
-                            s3?: {
-                                /**
-                                 * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                                 * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                                 * the destination bucket and prefix. The full list of possible canned acl can be found at
-                                 * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                                 * Please also note that by default only the object owner gets full controls. If you are using cross account
-                                 * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                                 * read the logs.
-                                 */
-                                canned_acl?: string;
-                                /**
-                                 * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                                 * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                                 * destination. Please also note that you cannot use AWS keys to deliver logs.
-                                 */
-                                destination?: string;
-                                /**
-                                 * (Optional) Flag to enable server side encryption, `false` by default.
-                                 */
-                                enable_encryption?: boolean;
-                                /**
-                                 * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                                 * encryption is enabled and the default type is `sse-s3`.
-                                 */
-                                encryption_type?: string;
-                                /**
-                                 * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                                 * If both are set, endpoint will be used.
-                                 */
-                                endpoint?: string;
-                                /**
-                                 * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                                 */
-                                kms_key?: string;
-                                /**
-                                 * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                                 * endpoint will be used.
-                                 */
-                                region?: string;
-                            };
-                            volumes?: {
-                                destination?: string;
-                            };
-                            /**
-                             * destination needs to be provided. e.g.
-                             * `{ "workspace" : { "destination" : "/Users/user1@databricks.com/my-init.sh" } }`
-                             */
-                            workspace?: {
-                                /**
-                                 * workspace files destination, e.g. `/Users/user1@databricks.com/my-init.sh`
-                                 */
-                                destination?: string;
-                            };
-                        }[];
-                        /**
-                         * The optional ID of the instance pool to which the cluster belongs.
-                         */
-                        instance_pool_id?: string;
-                        /**
-                         * This field encodes, through a single value, the resources available to each of
-                         * the Spark nodes in this cluster. For example, the Spark nodes can be provisioned
-                         * and optimized for memory or compute intensive workloads. A list of available node
-                         * types can be retrieved by using the :method:clusters/listNodeTypes API call.
-                         *
-                         */
-                        node_type_id?: string;
-                        /**
-                         * Number of worker nodes that this cluster should have. A cluster has one Spark Driver
-                         * and `num_workers` Executors for a total of `num_workers` + 1 Spark nodes.
-                         *
-                         * Note: When reading the properties of a cluster, this field reflects the desired number
-                         * of workers rather than the actual current number of workers. For instance, if a cluster
-                         * is resized from 5 to 10 workers, this field will immediately be updated to reflect
-                         * the target size of 10 workers, whereas the workers listed in `spark_info` will gradually
-                         * increase from 5 to 10 as the new nodes are provisioned.
-                         */
-                        num_workers?: number;
-                        /**
-                         * The ID of the cluster policy used to create the cluster if applicable.
-                         */
-                        policy_id?: string;
-                        runtime_engine?: string;
-                        /**
-                         * Single user name if data_security_mode is `SINGLE_USER`
-                         */
-                        single_user_name?: string;
-                        /**
-                         * An object containing a set of optional, user-specified Spark configuration key-value pairs.
-                         * Users can also pass in a string of extra JVM options to the driver and the executors via
-                         * `spark.driver.extraJavaOptions` and `spark.executor.extraJavaOptions` respectively.
-                         *
-                         */
-                        spark_conf?: {
-                            [k: string]: string;
-                        };
-                        /**
-                         * An object containing a set of optional, user-specified environment variable key-value pairs.
-                         * Please note that key-value pair of the form (X,Y) will be exported as is (i.e.,
-                         * `export X='Y'`) while launching the driver and workers.
-                         *
-                         * In order to specify an additional set of `SPARK_DAEMON_JAVA_OPTS`, we recommend appending
-                         * them to `$SPARK_DAEMON_JAVA_OPTS` as shown in the example below. This ensures that all
-                         * default databricks managed environmental variables are included as well.
-                         *
-                         * Example Spark environment variables:
-                         * `{"SPARK_WORKER_MEMORY": "28000m", "SPARK_LOCAL_DIRS": "/local_disk0"}` or
-                         * `{"SPARK_DAEMON_JAVA_OPTS": "$SPARK_DAEMON_JAVA_OPTS -Dspark.shuffle.service.enabled=true"}`
-                         */
-                        spark_env_vars?: {
-                            [k: string]: string;
-                        };
-                        /**
-                         * The Spark version of the cluster, e.g. `3.3.x-scala2.11`.
-                         * A list of available Spark versions can be retrieved by using
-                         * the :method:clusters/sparkVersions API call.
-                         *
-                         */
-                        spark_version?: string;
-                        /**
-                         * SSH public key contents that will be added to each Spark node in this cluster. The
-                         * corresponding private keys can be used to login with the user name `ubuntu` on port `2200`.
-                         * Up to 10 keys can be specified.
-                         */
-                        ssh_public_keys?: string[];
-                        workload_type?: {
-                            /**
-                             *  defined what type of clients can use the cluster. E.g. Notebooks, Jobs
-                             */
-                            clients?: {
-                                /**
-                                 * With jobs set, the cluster can be used for jobs
-                                 */
-                                jobs?: boolean;
-                                /**
-                                 * With notebooks set, this cluster can be used for notebooks
-                                 */
-                                notebooks?: boolean;
-                            };
-                        };
-                    };
-                }[];
-                /**
-                 * An optional maximum allowed number of concurrent runs of the job.
-                 *
-                 * Set this value if you want to be able to execute multiple runs of the same job concurrently. This is useful for example if you trigger your job on a frequent schedule and want to allow consecutive runs to overlap with each other, or if you want to trigger multiple runs which differ by their input parameters.
-                 *
-                 * This setting affects only new runs. For example, suppose the job’s concurrency is 4 and there are 4 concurrent active runs. Then setting the concurrency to 3 won’t kill any of the active runs. However, from then on, new runs are skipped unless there are fewer than 3 active runs.
-                 *
-                 * This value cannot exceed 1000\. Setting this value to 0 causes all new runs to be skipped. The default behavior is to allow only 1 concurrent run.
-                 */
-                max_concurrent_runs?: number;
-                /**
-                 * An optional name for the job.
-                 */
-                name?: string;
-                /**
-                 * Optional notification settings that are used when sending notifications to each of the `email_notifications` and `webhook_notifications` for this job.
-                 */
-                notification_settings?: {
-                    /**
-                     * If true, do not send notifications to recipients specified in `on_failure` if the run is canceled.
-                     */
-                    no_alert_for_canceled_runs?: boolean;
-                    /**
-                     * If true, do not send notifications to recipients specified in `on_failure` if the run is skipped.
-                     */
-                    no_alert_for_skipped_runs?: boolean;
-                };
-                /**
-                 * Job-level parameter definitions
-                 */
-                parameters?: {
-                    /**
-                     * Default value of the parameter.
-                     */
-                    default: string;
-                    /**
-                     * The name of the defined parameter. May only contain alphanumeric characters, `_`, `-`, and `.`
-                     */
-                    name: string;
-                }[];
-                permissions?: {
-                    group_name?: string;
-                    level: string;
-                    service_principal_name?: string;
-                    user_name?: string;
-                }[];
-                run_as?: {
-                    /**
-                     * Application ID of an active service principal. Setting this field requires the `servicePrincipal/user` role.
-                     */
-                    service_principal_name?: string;
-                    /**
-                     * The email of an active workspace user. Non-admin users can only set this field to their own email.
-                     */
-                    user_name?: string;
-                };
-                /**
-                 * An optional periodic schedule for this job. The default behavior is that the job only runs when triggered by clicking “Run Now” in the Jobs UI or sending an API request to `runNow`.
-                 */
-                schedule?: {
-                    /**
-                     * Whether this trigger is paused or not.
-                     */
-                    pause_status?: string;
-                    /**
-                     * A Cron expression using Quartz syntax that describes the schedule for a job.
-                     * See [Cron Trigger](http://www.quartz-scheduler.org/documentation/quartz-2.3.0/tutorials/crontrigger.html)
-                     * for details. This field is required."
-                     *
-                     */
-                    quartz_cron_expression: string;
-                    /**
-                     * A Java timezone ID. The schedule for a job is resolved with respect to this timezone.
-                     * See [Java TimeZone](https://docs.oracle.com/javase/7/docs/api/java/util/TimeZone.html) for details.
-                     * This field is required.
-                     *
-                     */
-                    timezone_id: string;
-                };
-                /**
-                 * A map of tags associated with the job. These are forwarded to the cluster as cluster tags for jobs clusters, and are subject to the same limitations as cluster tags. A maximum of 25 tags can be added to the job.
-                 */
-                tags?: {
-                    [k: string]: string;
-                };
-                /**
-                 * A list of task specifications to be executed by this job.
-                 */
-                tasks?: {
-                    /**
-                     * The key of the compute requirement, specified in `job.settings.compute`, to use for execution of this task.
-                     */
-                    compute_key?: string;
-                    /**
-                     * If condition_task, specifies a condition with an outcome that can be used to control the execution of other tasks. Does not require a cluster to execute and does not support retries or notifications.
-                     */
-                    condition_task?: {
-                        /**
-                         * The left operand of the condition task. Can be either a string value or a job state or parameter reference.
-                         */
-                        left?: string;
-                        /**
-                         * * `EQUAL_TO`, `NOT_EQUAL` operators perform string comparison of their operands. This means that `“12.0” == “12”` will evaluate to `false`.
-                         * * `GREATER_THAN`, `GREATER_THAN_OR_EQUAL`, `LESS_THAN`, `LESS_THAN_OR_EQUAL` operators perform numeric comparison of their operands. `“12.0” >= “12”` will evaluate to `true`, `“10.0” >= “12”` will evaluate to `false`.
-                         *
-                         * The boolean comparison to task values can be implemented with operators `EQUAL_TO`, `NOT_EQUAL`. If a task value was set to a boolean value, it will be serialized to `“true”` or `“false”` for the comparison.
-                         *
-                         */
-                        op?: string;
-                        /**
-                         * The right operand of the condition task. Can be either a string value or a job state or parameter reference.
-                         */
-                        right?: string;
-                    };
-                    /**
-                     * If dbt_task, indicates that this must execute a dbt task. It requires both Databricks SQL and the ability to use a serverless or a pro SQL warehouse.
-                     */
-                    dbt_task?: {
-                        /**
-                         * Optional name of the catalog to use. The value is the top level in the 3-level namespace of Unity Catalog (catalog / schema / relation). The catalog value can only be specified if a warehouse_id is specified. Requires dbt-databricks >= 1.1.1.
-                         */
-                        catalog?: string;
-                        /**
-                         * A list of dbt commands to execute. All commands must start with `dbt`. This parameter must not be empty. A maximum of up to 10 commands can be provided.
-                         */
-                        commands: string[];
-                        /**
-                         * Optional (relative) path to the profiles directory. Can only be specified if no warehouse_id is specified. If no warehouse_id is specified and this folder is unset, the root directory is used.
-                         */
-                        profiles_directory?: string;
-                        /**
-                         * Optional (relative) path to the project directory, if no value is provided, the root of the git repository is used.
-                         */
-                        project_directory?: string;
-                        /**
-                         * Optional schema to write to. This parameter is only used when a warehouse_id is also provided. If not provided, the `default` schema is used.
-                         */
-                        schema?: string;
-                        /**
-                         * ID of the SQL warehouse to connect to. If provided, we automatically generate and provide the profile and connection details to dbt. It can be overridden on a per-command basis by using the `--profiles-dir` command line argument.
-                         */
-                        warehouse_id?: string;
-                    };
-                    /**
-                     * An optional array of objects specifying the dependency graph of the task. All tasks specified in this field must complete successfully before executing this task.
-                     * The key is `task_key`, and the value is the name assigned to the dependent task.
-                     *
-                     */
-                    depends_on?: {
-                        /**
-                         * Can only be specified on condition task dependencies. The outcome of the dependent task that must be met for this task to run.
-                         */
-                        outcome?: string;
-                        /**
-                         * The name of the task this task depends on.
-                         */
-                        task_key: string;
-                    }[];
-                    /**
-                     * An optional description for this task.
-                     * The maximum length is 4096 bytes.
-                     */
-                    description?: string;
-                    /**
-                     * An optional set of email addresses that is notified when runs of this task begin or complete as well as when this task is deleted. The default behavior is to not send any emails.
-                     */
-                    email_notifications?: {
-                        on_duration_warning_threshold_exceeded?: string[];
-                        /**
-                         * A list of email addresses to be notified when a run unsuccessfully completes. A run is considered to have completed unsuccessfully if it ends with an `INTERNAL_ERROR` `life_cycle_state` or a `FAILED`, or `TIMED_OUT` result_state. If this is not specified on job creation, reset, or update the list is empty, and notifications are not sent.
-                         */
-                        on_failure?: string[];
-                        /**
-                         * A list of email addresses to be notified when a run begins. If not specified on job creation, reset, or update, the list is empty, and notifications are not sent.
-                         */
-                        on_start?: string[];
-                        /**
-                         * A list of email addresses to be notified when a run successfully completes. A run is considered to have completed successfully if it ends with a `TERMINATED` `life_cycle_state` and a `SUCCESS` result_state. If not specified on job creation, reset, or update, the list is empty, and notifications are not sent.
-                         */
-                        on_success?: string[];
-                    };
-                    /**
-                     * If existing_cluster_id, the ID of an existing cluster that is used for all runs of this task. When running tasks on an existing cluster, you may need to manually restart the cluster if it stops responding. We suggest running jobs on new clusters for greater reliability.
-                     */
-                    existing_cluster_id?: string;
-                    health?: {
-                        rules?: {
-                            metric?: string;
-                            op?: string;
-                            value?: number;
-                        }[];
-                    };
-                    /**
-                     * If job_cluster_key, this task is executed reusing the cluster specified in `job.settings.job_clusters`.
-                     */
-                    job_cluster_key?: string;
-                    /**
-                     * An optional list of libraries to be installed on the cluster that executes the task. The default value is an empty list.
-                     */
-                    libraries?: {
-                        /**
-                         * Specification of a CRAN library to be installed as part of the library
-                         */
-                        cran?: {
-                            /**
-                             * The name of the CRAN package to install.
-                             */
-                            package: string;
-                            /**
-                             * The repository where the package can be found. If not specified, the default CRAN repo is used.
-                             */
-                            repo?: string;
-                        };
-                        /**
-                         * URI of the egg to be installed. Currently only DBFS and S3 URIs are supported.
-                         * For example: `{ "egg": "dbfs:/my/egg" }` or
-                         * `{ "egg": "s3://my-bucket/egg" }`.
-                         * If S3 is used, please make sure the cluster has read access on the library. You may need to
-                         * launch the cluster with an IAM role to access the S3 URI.
-                         */
-                        egg?: string;
-                        /**
-                         * URI of the jar to be installed. Currently only DBFS and S3 URIs are supported.
-                         * For example: `{ "jar": "dbfs:/mnt/databricks/library.jar" }` or
-                         * `{ "jar": "s3://my-bucket/library.jar" }`.
-                         * If S3 is used, please make sure the cluster has read access on the library. You may need to
-                         * launch the cluster with an IAM role to access the S3 URI.
-                         */
-                        jar?: string;
-                        /**
-                         * Specification of a maven library to be installed. For example:
-                         * `{ "coordinates": "org.jsoup:jsoup:1.7.2" }`
-                         */
-                        maven?: {
-                            /**
-                             * Gradle-style maven coordinates. For example: "org.jsoup:jsoup:1.7.2".
-                             */
-                            coordinates: string;
-                            /**
-                             * List of dependences to exclude. For example: `["slf4j:slf4j", "*:hadoop-client"]`.
-                             *
-                             * Maven dependency exclusions:
-                             * https://maven.apache.org/guides/introduction/introduction-to-optional-and-excludes-dependencies.html.
-                             */
-                            exclusions?: string[];
-                            /**
-                             * Maven repo to install the Maven package from. If omitted, both Maven Central Repository
-                             * and Spark Packages are searched.
-                             */
-                            repo?: string;
-                        };
-                        /**
-                         * Specification of a PyPi library to be installed. For example:
-                         * `{ "package": "simplejson" }`
-                         */
-                        pypi?: {
-                            /**
-                             * The name of the pypi package to install. An optional exact version specification is also
-                             * supported. Examples: "simplejson" and "simplejson==3.8.0".
-                             */
-                            package: string;
-                            /**
-                             * The repository where the package can be found. If not specified, the default pip index is
-                             * used.
-                             */
-                            repo?: string;
-                        };
-                        /**
-                         * URI of the wheel to be installed.
-                         * For example: `{ "whl": "dbfs:/my/whl" }` or `{ "whl": "s3://my-bucket/whl" }`.
-                         * If S3 is used, please make sure the cluster has read access on the library. You may need to
-                         * launch the cluster with an IAM role to access the S3 URI.
-                         */
-                        whl?: string;
-                    }[];
-                    /**
-                     * An optional maximum number of times to retry an unsuccessful run. A run is considered to be unsuccessful if it completes with the `FAILED` result_state or `INTERNAL_ERROR` `life_cycle_state`. The value -1 means to retry indefinitely and the value 0 means to never retry. The default behavior is to never retry.
-                     */
-                    max_retries?: number;
-                    /**
-                     * An optional minimal interval in milliseconds between the start of the failed run and the subsequent retry run. The default behavior is that unsuccessful runs are immediately retried.
-                     */
-                    min_retry_interval_millis?: number;
-                    /**
-                     * If new_cluster, a description of a cluster that is created for only for this task.
-                     */
-                    new_cluster?: {
-                        /**
-                         * Parameters needed in order to automatically scale clusters up and down based on load.
-                         * Note: autoscaling works best with DB runtime versions 3.0 or later.
-                         */
-                        autoscale?: {
-                            /**
-                             * The maximum number of workers to which the cluster can scale up when overloaded.
-                             * Note that `max_workers` must be strictly greater than `min_workers`.
-                             */
-                            max_workers: number;
-                            /**
-                             * The minimum number of workers to which the cluster can scale down when underutilized.
-                             * It is also the initial number of workers the cluster will have after creation.
-                             */
-                            min_workers: number;
-                        };
-                        /**
-                         * Automatically terminates the cluster after it is inactive for this time in minutes. If not set,
-                         * this cluster will not be automatically terminated. If specified, the threshold must be between
-                         * 10 and 10000 minutes.
-                         * Users can also set this value to 0 to explicitly disable automatic termination.
-                         */
-                        autotermination_minutes?: number;
-                        /**
-                         * Attributes related to clusters running on Amazon Web Services.
-                         * If not specified at cluster creation, a set of default values will be used.
-                         */
-                        aws_attributes?: {
-                            availability?: string;
-                            /**
-                             * The number of volumes launched for each instance. Users can choose up to 10 volumes.
-                             * This feature is only enabled for supported node types. Legacy node types cannot specify
-                             * custom EBS volumes.
-                             * For node types with no instance store, at least one EBS volume needs to be specified;
-                             * otherwise, cluster creation will fail.
-                             *
-                             * These EBS volumes will be mounted at `/ebs0`, `/ebs1`, and etc.
-                             * Instance store volumes will be mounted at `/local_disk0`, `/local_disk1`, and etc.
-                             *
-                             * If EBS volumes are attached, Databricks will configure Spark to use only the EBS volumes for
-                             * scratch storage because heterogenously sized scratch devices can lead to inefficient disk
-                             * utilization. If no EBS volumes are attached, Databricks will configure Spark to use instance
-                             * store volumes.
-                             *
-                             * Please note that if EBS volumes are specified, then the Spark configuration `spark.local.dir`
-                             * will be overridden.
-                             */
-                            ebs_volume_count?: number;
-                            /**
-                             * <needs content added>
-                             */
-                            ebs_volume_iops?: number;
-                            /**
-                             * The size of each EBS volume (in GiB) launched for each instance. For general purpose
-                             * SSD, this value must be within the range 100 - 4096. For throughput optimized HDD,
-                             * this value must be within the range 500 - 4096.
-                             */
-                            ebs_volume_size?: number;
-                            /**
-                             * <needs content added>
-                             */
-                            ebs_volume_throughput?: number;
-                            ebs_volume_type?: string;
-                            /**
-                             * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                             * If this value is greater than 0, the cluster driver node in particular will be placed on an
-                             * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                             * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                             * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                             * be placed on `availability` instances. Note that this value does not affect
-                             * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                             */
-                            first_on_demand?: number;
-                            /**
-                             * Nodes for this cluster will only be placed on AWS instances with this instance profile. If
-                             * ommitted, nodes will be placed on instances without an IAM instance profile. The instance
-                             * profile must have previously been added to the Databricks environment by an account
-                             * administrator.
-                             *
-                             * This feature may only be available to certain customer plans.
-                             *
-                             * If this field is ommitted, we will pull in the default from the conf if it exists.
-                             */
-                            instance_profile_arn?: string;
-                            /**
-                             * The bid price for AWS spot instances, as a percentage of the corresponding instance type's
-                             * on-demand price.
-                             * For example, if this field is set to 50, and the cluster needs a new `r3.xlarge` spot
-                             * instance, then the bid price is half of the price of
-                             * on-demand `r3.xlarge` instances. Similarly, if this field is set to 200, the bid price is twice
-                             * the price of on-demand `r3.xlarge` instances. If not specified, the default value is 100.
-                             * When spot instances are requested for this cluster, only spot instances whose bid price
-                             * percentage matches this field will be considered.
-                             * Note that, for safety, we enforce this field to be no more than 10000.
-                             *
-                             * The default value and documentation here should be kept consistent with
-                             * CommonConf.defaultSpotBidPricePercent and CommonConf.maxSpotBidPricePercent.
-                             */
-                            spot_bid_price_percent?: number;
-                            /**
-                             * Identifier for the availability zone/datacenter in which the cluster resides.
-                             * This string will be of a form like "us-west-2a". The provided availability
-                             * zone must be in the same region as the Databricks deployment. For example, "us-west-2a"
-                             * is not a valid zone id if the Databricks deployment resides in the "us-east-1" region.
-                             * This is an optional field at cluster creation, and if not specified, a default zone will be used.
-                             * If the zone specified is "auto", will try to place cluster in a zone with high availability,
-                             * and will retry placement in a different AZ if there is not enough capacity.
-                             * See [[AutoAZHelper.scala]] for more details.
-                             * The list of available zones as well as the default value can be found by using the
-                             * `List Zones`_ method.
-                             */
-                            zone_id?: string;
-                        };
-                        /**
-                         * Attributes related to clusters running on Microsoft Azure.
-                         * If not specified at cluster creation, a set of default values will be used.
-                         */
-                        azure_attributes?: {
-                            availability?: string;
-                            /**
-                             * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                             * This value should be greater than 0, to make sure the cluster driver node is placed on an
-                             * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                             * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                             * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                             * be placed on `availability` instances. Note that this value does not affect
-                             * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                             */
-                            first_on_demand?: number;
-                            /**
-                             * Defines values necessary to configure and run Azure Log Analytics agent
-                             */
-                            log_analytics_info?: {
-                                /**
-                                 * <needs content added>
-                                 */
-                                log_analytics_primary_key?: string;
-                                /**
-                                 * <needs content added>
-                                 */
-                                log_analytics_workspace_id?: string;
-                            };
-                            /**
-                             * The max bid price to be used for Azure spot instances.
-                             * The Max price for the bid cannot be higher than the on-demand price of the instance.
-                             * If not specified, the default value is -1, which specifies that the instance cannot be evicted
-                             * on the basis of price, and only on the basis of availability. Further, the value should > 0 or -1.
-                             */
-                            spot_bid_max_price?: number;
-                        };
-                        /**
-                         * The configuration for delivering spark logs to a long-term storage destination.
-                         * Two kinds of destinations (dbfs and s3) are supported. Only one destination can be specified
-                         * for one cluster. If the conf is given, the logs will be delivered to the destination every
-                         * `5 mins`. The destination of driver logs is `$destination/$clusterId/driver`, while
-                         * the destination of executor logs is `$destination/$clusterId/executor`.
-                         */
-                        cluster_log_conf?: {
-                            /**
-                             * destination needs to be provided. e.g.
-                             * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                             */
-                            dbfs?: {
-                                /**
-                                 * dbfs destination, e.g. `dbfs:/my/path`
-                                 */
-                                destination?: string;
-                            };
-                            /**
-                             * destination and either the region or endpoint need to be provided. e.g.
-                             * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                             * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                             * `instance_profile_arn` has permission to write data to the s3 destination.
-                             */
-                            s3?: {
-                                /**
-                                 * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                                 * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                                 * the destination bucket and prefix. The full list of possible canned acl can be found at
-                                 * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                                 * Please also note that by default only the object owner gets full controls. If you are using cross account
-                                 * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                                 * read the logs.
-                                 */
-                                canned_acl?: string;
-                                /**
-                                 * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                                 * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                                 * destination. Please also note that you cannot use AWS keys to deliver logs.
-                                 */
-                                destination?: string;
-                                /**
-                                 * (Optional) Flag to enable server side encryption, `false` by default.
-                                 */
-                                enable_encryption?: boolean;
-                                /**
-                                 * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                                 * encryption is enabled and the default type is `sse-s3`.
-                                 */
-                                encryption_type?: string;
-                                /**
-                                 * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                                 * If both are set, endpoint will be used.
-                                 */
-                                endpoint?: string;
-                                /**
-                                 * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                                 */
-                                kms_key?: string;
-                                /**
-                                 * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                                 * endpoint will be used.
-                                 */
-                                region?: string;
-                            };
-                        };
-                        /**
-                         * Cluster name requested by the user. This doesn't have to be unique.
-                         * If not specified at creation, the cluster name will be an empty string.
-                         *
-                         */
-                        cluster_name?: string;
-                        cluster_source?: string;
-                        /**
-                         * Additional tags for cluster resources. Databricks will tag all cluster resources (e.g., AWS
-                         * instances and EBS volumes) with these tags in addition to `default_tags`. Notes:
-                         *
-                         * - Currently, Databricks allows at most 45 custom tags
-                         *
-                         * - Clusters can only reuse cloud resources if the resources' tags are a subset of the cluster tags
-                         */
-                        custom_tags?: {
-                            [k: string]: string;
-                        };
-                        data_security_mode?: string;
-                        docker_image?: {
-                            basic_auth?: {
-                                /**
-                                 * Password of the user
-                                 */
-                                password?: string;
-                                /**
-                                 * Name of the user
-                                 */
-                                username?: string;
-                            };
-                            /**
-                             * URL of the docker image.
-                             */
-                            url?: string;
-                        };
-                        /**
-                         * The optional ID of the instance pool for the driver of the cluster belongs.
-                         * The pool cluster uses the instance pool with id (instance_pool_id) if the driver pool is not
-                         * assigned.
-                         */
-                        driver_instance_pool_id?: string;
-                        /**
-                         * The node type of the Spark driver. Note that this field is optional;
-                         * if unset, the driver node type will be set as the same value
-                         * as `node_type_id` defined above.
-                         *
-                         */
-                        driver_node_type_id?: string;
-                        /**
-                         * Autoscaling Local Storage: when enabled, this cluster will dynamically acquire additional disk
-                         * space when its Spark workers are running low on disk space. This feature requires specific AWS
-                         * permissions to function correctly - refer to the User Guide for more details.
-                         */
-                        enable_elastic_disk?: boolean;
-                        /**
-                         * Whether to enable LUKS on cluster VMs' local disks
-                         */
-                        enable_local_disk_encryption?: boolean;
-                        /**
-                         * Attributes related to clusters running on Google Cloud Platform.
-                         * If not specified at cluster creation, a set of default values will be used.
-                         */
-                        gcp_attributes?: {
-                            availability?: string;
-                            /**
-                             * boot disk size in GB
-                             */
-                            boot_disk_size?: number;
-                            /**
-                             * If provided, the cluster will impersonate the google service account when accessing
-                             * gcloud services (like GCS). The google service account
-                             * must have previously been added to the Databricks environment by an account
-                             * administrator.
-                             */
-                            google_service_account?: string;
-                            /**
-                             * If provided, each node (workers and driver) in the cluster will have this number of local SSDs attached. Each local SSD is 375GB in size. Refer to [GCP documentation](https://cloud.google.com/compute/docs/disks/local-ssd#choose_number_local_ssds) for the supported number of local SSDs for each instance type.
-                             */
-                            local_ssd_count?: number;
-                        };
-                        /**
-                         * The configuration for storing init scripts. Any number of destinations can be specified. The scripts are executed sequentially in the order provided. If `cluster_log_conf` is specified, init script logs are sent to `<destination>/<cluster-ID>/init_scripts`.
-                         */
-                        init_scripts?: {
-                            /**
-                             * destination needs to be provided. e.g.
-                             * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                             */
-                            dbfs?: {
-                                /**
-                                 * dbfs destination, e.g. `dbfs:/my/path`
-                                 */
-                                destination?: string;
-                            };
-                            /**
-                             * destination and either the region or endpoint need to be provided. e.g.
-                             * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                             * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                             * `instance_profile_arn` has permission to write data to the s3 destination.
-                             */
-                            s3?: {
-                                /**
-                                 * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                                 * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                                 * the destination bucket and prefix. The full list of possible canned acl can be found at
-                                 * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                                 * Please also note that by default only the object owner gets full controls. If you are using cross account
-                                 * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                                 * read the logs.
-                                 */
-                                canned_acl?: string;
-                                /**
-                                 * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                                 * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                                 * destination. Please also note that you cannot use AWS keys to deliver logs.
-                                 */
-                                destination?: string;
-                                /**
-                                 * (Optional) Flag to enable server side encryption, `false` by default.
-                                 */
-                                enable_encryption?: boolean;
-                                /**
-                                 * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                                 * encryption is enabled and the default type is `sse-s3`.
-                                 */
-                                encryption_type?: string;
-                                /**
-                                 * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                                 * If both are set, endpoint will be used.
-                                 */
-                                endpoint?: string;
-                                /**
-                                 * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                                 */
-                                kms_key?: string;
-                                /**
-                                 * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                                 * endpoint will be used.
-                                 */
-                                region?: string;
-                            };
-                            volumes?: {
-                                destination?: string;
-                            };
-                            /**
-                             * destination needs to be provided. e.g.
-                             * `{ "workspace" : { "destination" : "/Users/user1@databricks.com/my-init.sh" } }`
-                             */
-                            workspace?: {
-                                /**
-                                 * workspace files destination, e.g. `/Users/user1@databricks.com/my-init.sh`
-                                 */
-                                destination?: string;
-                            };
-                        }[];
-                        /**
-                         * The optional ID of the instance pool to which the cluster belongs.
-                         */
-                        instance_pool_id?: string;
-                        /**
-                         * This field encodes, through a single value, the resources available to each of
-                         * the Spark nodes in this cluster. For example, the Spark nodes can be provisioned
-                         * and optimized for memory or compute intensive workloads. A list of available node
-                         * types can be retrieved by using the :method:clusters/listNodeTypes API call.
-                         *
-                         */
-                        node_type_id?: string;
-                        /**
-                         * Number of worker nodes that this cluster should have. A cluster has one Spark Driver
-                         * and `num_workers` Executors for a total of `num_workers` + 1 Spark nodes.
-                         *
-                         * Note: When reading the properties of a cluster, this field reflects the desired number
-                         * of workers rather than the actual current number of workers. For instance, if a cluster
-                         * is resized from 5 to 10 workers, this field will immediately be updated to reflect
-                         * the target size of 10 workers, whereas the workers listed in `spark_info` will gradually
-                         * increase from 5 to 10 as the new nodes are provisioned.
-                         */
-                        num_workers?: number;
-                        /**
-                         * The ID of the cluster policy used to create the cluster if applicable.
-                         */
-                        policy_id?: string;
-                        runtime_engine?: string;
-                        /**
-                         * Single user name if data_security_mode is `SINGLE_USER`
-                         */
-                        single_user_name?: string;
-                        /**
-                         * An object containing a set of optional, user-specified Spark configuration key-value pairs.
-                         * Users can also pass in a string of extra JVM options to the driver and the executors via
-                         * `spark.driver.extraJavaOptions` and `spark.executor.extraJavaOptions` respectively.
-                         *
-                         */
-                        spark_conf?: {
-                            [k: string]: string;
-                        };
-                        /**
-                         * An object containing a set of optional, user-specified environment variable key-value pairs.
-                         * Please note that key-value pair of the form (X,Y) will be exported as is (i.e.,
-                         * `export X='Y'`) while launching the driver and workers.
-                         *
-                         * In order to specify an additional set of `SPARK_DAEMON_JAVA_OPTS`, we recommend appending
-                         * them to `$SPARK_DAEMON_JAVA_OPTS` as shown in the example below. This ensures that all
-                         * default databricks managed environmental variables are included as well.
-                         *
-                         * Example Spark environment variables:
-                         * `{"SPARK_WORKER_MEMORY": "28000m", "SPARK_LOCAL_DIRS": "/local_disk0"}` or
-                         * `{"SPARK_DAEMON_JAVA_OPTS": "$SPARK_DAEMON_JAVA_OPTS -Dspark.shuffle.service.enabled=true"}`
-                         */
-                        spark_env_vars?: {
-                            [k: string]: string;
-                        };
-                        /**
-                         * The Spark version of the cluster, e.g. `3.3.x-scala2.11`.
-                         * A list of available Spark versions can be retrieved by using
-                         * the :method:clusters/sparkVersions API call.
-                         *
-                         */
-                        spark_version?: string;
-                        /**
-                         * SSH public key contents that will be added to each Spark node in this cluster. The
-                         * corresponding private keys can be used to login with the user name `ubuntu` on port `2200`.
-                         * Up to 10 keys can be specified.
-                         */
-                        ssh_public_keys?: string[];
-                        workload_type?: {
-                            /**
-                             *  defined what type of clients can use the cluster. E.g. Notebooks, Jobs
-                             */
-                            clients?: {
-                                /**
-                                 * With jobs set, the cluster can be used for jobs
-                                 */
-                                jobs?: boolean;
-                                /**
-                                 * With notebooks set, this cluster can be used for notebooks
-                                 */
-                                notebooks?: boolean;
-                            };
-                        };
-                    };
-                    /**
-                     * If notebook_task, indicates that this task must run a notebook. This field may not be specified in conjunction with spark_jar_task.
-                     */
-                    notebook_task?: {
-                        /**
-                         * Base parameters to be used for each run of this job. If the run is initiated by a call to
-                         * :method:jobs/runNow with parameters specified, the two parameters maps are merged. If the same key is specified in
-                         * `base_parameters` and in `run-now`, the value from `run-now` is used.
-                         *
-                         * Use [Task parameter variables](https://docs.databricks.com/jobs.html#parameter-variables) to set parameters containing information about job runs.
-                         *
-                         * If the notebook takes a parameter that is not specified in the job’s `base_parameters` or the `run-now` override parameters,
-                         * the default value from the notebook is used.
-                         *
-                         * Retrieve these parameters in a notebook using [dbutils.widgets.get](https://docs.databricks.com/dev-tools/databricks-utils.html#dbutils-widgets).
-                         *
-                         */
-                        base_parameters?: {
-                            [k: string]: string;
-                        };
-                        /**
-                         * The path of the notebook to be run in the Databricks workspace or remote repository.
-                         * For notebooks stored in the Databricks workspace, the path must be absolute and begin with a slash.
-                         * For notebooks stored in a remote repository, the path must be relative. This field is required.
-                         *
-                         */
-                        notebook_path: string;
-                        /**
-                         * Optional location type of the Python file. When set to `WORKSPACE` or not specified, the file will be retrieved
-                         * from the local <Databricks> workspace or cloud location (if the `python_file` has a URI format). When set to `GIT`,
-                         * the Python file will be retrieved from a Git repository defined in `git_source`.
-                         *
-                         * * `WORKSPACE`: The Python file is located in a <Databricks> workspace or at a cloud filesystem URI.
-                         * * `GIT`: The Python file is located in a remote Git repository.
-                         *
-                         */
-                        source?: string;
-                    };
-                    /**
-                     * Optional notification settings that are used when sending notifications to each of the `email_notifications` for this task.
-                     */
-                    notification_settings?: {
-                        /**
-                         * If true, do not send notifications to recipients specified in `on_start` for the retried runs and do not send notifications to recipients specified in `on_failure` until the last retry of the run.
-                         */
-                        alert_on_last_attempt?: boolean;
-                        /**
-                         * If true, do not send notifications to recipients specified in `on_failure` if the run is canceled.
-                         */
-                        no_alert_for_canceled_runs?: boolean;
-                        /**
-                         * If true, do not send notifications to recipients specified in `on_failure` if the run is skipped.
-                         */
-                        no_alert_for_skipped_runs?: boolean;
-                    };
-                    /**
-                     * If pipeline_task, indicates that this task must execute a Pipeline.
-                     */
-                    pipeline_task?: {
-                        /**
-                         * If true, a full refresh will be triggered on the delta live table.
-                         */
-                        full_refresh?: boolean;
-                        /**
-                         * The full name of the pipeline task to execute.
-                         */
-                        pipeline_id?: string;
-                    };
-                    /**
-                     * If python_wheel_task, indicates that this job must execute a PythonWheel.
-                     */
-                    python_wheel_task?: {
-                        /**
-                         * Named entry point to use, if it does not exist in the metadata of the package it executes the function from the package directly using `$packageName.$entryPoint()`
-                         */
-                        entry_point?: string;
-                        /**
-                         * Command-line parameters passed to Python wheel task in the form of `["--name=task", "--data=dbfs:/path/to/data.json"]`. Leave it empty if `parameters` is not null.
-                         */
-                        named_parameters?: {
-                            [k: string]: string;
-                        };
-                        /**
-                         * Name of the package to execute
-                         */
-                        package_name?: string;
-                        /**
-                         * Command-line parameters passed to Python wheel task. Leave it empty if `named_parameters` is not null.
-                         */
-                        parameters?: string[];
-                    };
-                    /**
-                     * An optional policy to specify whether to retry a task when it times out. The default behavior is to not retry on timeout.
-                     */
-                    retry_on_timeout?: boolean;
-                    /**
-                     * An optional value specifying the condition determining whether the task is run once its dependencies have been completed. When omitted, defaults to `ALL_SUCCESS`.
-                     *
-                     * * `ALL_SUCCESS`: All dependencies have executed and succeeded
-                     * * `AT_LEAST_ONE_SUCCESS`: At least one dependency has succeeded
-                     * * `NONE_FAILED`: None of the dependencies have failed and at least one was executed
-                     * * `ALL_DONE`: All dependencies completed and at least one was executed
-                     * * `AT_LEAST_ONE_FAILED`: At least one dependency failed
-                     * * `ALL_FAILED`: ALl dependencies have failed
-                     *
-                     */
-                    run_if?: string;
-                    run_job_task?: {
-                        job_id: number;
-                        job_parameters?: unknown;
-                    };
-                    /**
-                     * If spark_jar_task, indicates that this task must run a JAR.
-                     */
-                    spark_jar_task?: {
-                        /**
-                         * Deprecated since 04/2016. Provide a `jar` through the `libraries` field instead. For an example, see :method:jobs/create.
-                         *
-                         */
-                        jar_uri?: string;
-                        /**
-                         * The full name of the class containing the main method to be executed. This class must be contained in a JAR provided as a library.
-                         *
-                         * The code must use `SparkContext.getOrCreate` to obtain a Spark context; otherwise, runs of the job fail.
-                         */
-                        main_class_name?: string;
-                        /**
-                         * Parameters passed to the main method.
-                         *
-                         * Use [Task parameter variables](https://docs.databricks.com/jobs.html#parameter-variables) to set parameters containing information about job runs.
-                         *
-                         */
-                        parameters?: string[];
-                    };
-                    /**
-                     * If spark_python_task, indicates that this task must run a Python file.
-                     */
-                    spark_python_task?: {
-                        /**
-                         * Command line parameters passed to the Python file.
-                         *
-                         * Use [Task parameter variables](https://docs.databricks.com/jobs.html#parameter-variables) to set parameters containing information about job runs.
-                         *
-                         */
-                        parameters?: string[];
-                        /**
-                         * The Python file to be executed. Cloud file URIs (such as dbfs:/, s3:/, adls:/, gcs:/) and workspace paths are supported. For python files stored in the Databricks workspace, the path must be absolute and begin with `/`. For files stored in a remote repository, the path must be relative. This field is required.
-                         */
-                        python_file: string;
-                        /**
-                         * Optional location type of the Python file. When set to `WORKSPACE` or not specified, the file will be retrieved
-                         * from the local <Databricks> workspace or cloud location (if the `python_file` has a URI format). When set to `GIT`,
-                         * the Python file will be retrieved from a Git repository defined in `git_source`.
-                         *
-                         * * `WORKSPACE`: The Python file is located in a <Databricks> workspace or at a cloud filesystem URI.
-                         * * `GIT`: The Python file is located in a remote Git repository.
-                         *
-                         */
-                        source?: string;
-                    };
-                    /**
-                     * If spark_submit_task, indicates that this task must be launched by the spark submit script. This task can run only on new clusters.
-                     */
-                    spark_submit_task?: {
-                        /**
-                         * Command-line parameters passed to spark submit.
-                         *
-                         * Use [Task parameter variables](https://docs.databricks.com/jobs.html#parameter-variables) to set parameters containing information about job runs.
-                         *
-                         */
-                        parameters?: string[];
-                    };
-                    /**
-                     * If sql_task, indicates that this job must execute a SQL task.
-                     */
-                    sql_task?: {
-                        /**
-                         * If alert, indicates that this job must refresh a SQL alert.
-                         */
-                        alert?: {
-                            /**
-                             * The canonical identifier of the SQL alert.
-                             */
-                            alert_id: string;
-                            /**
-                             * If true, the alert notifications are not sent to subscribers.
-                             */
-                            pause_subscriptions?: boolean;
-                            /**
-                             * If specified, alert notifications are sent to subscribers.
-                             */
-                            subscriptions?: {
-                                /**
-                                 * The canonical identifier of the destination to receive email notification.
-                                 */
-                                destination_id?: string;
-                                /**
-                                 * The user name to receive the subscription email.
-                                 */
-                                user_name?: string;
-                            }[];
-                        };
-                        /**
-                         * If dashboard, indicates that this job must refresh a SQL dashboard.
-                         */
-                        dashboard?: {
-                            /**
-                             * Subject of the email sent to subscribers of this task.
-                             */
-                            custom_subject?: string;
-                            /**
-                             * The canonical identifier of the SQL dashboard.
-                             */
-                            dashboard_id: string;
-                            /**
-                             * If true, the dashboard snapshot is not taken, and emails are not sent to subscribers.
-                             */
-                            pause_subscriptions?: boolean;
-                            /**
-                             * If specified, dashboard snapshots are sent to subscriptions.
-                             */
-                            subscriptions?: {
-                                /**
-                                 * The canonical identifier of the destination to receive email notification.
-                                 */
-                                destination_id?: string;
-                                /**
-                                 * The user name to receive the subscription email.
-                                 */
-                                user_name?: string;
-                            }[];
-                        };
-                        /**
-                         * If file, indicates that this job runs a SQL file in a remote Git repository. Only one SQL statement is supported in a file. Multiple SQL statements separated by semicolons (;) are not permitted.
-                         */
-                        file?: {
-                            /**
-                             * Relative path of the SQL file in the remote Git repository.
-                             */
-                            path: string;
-                        };
-                        /**
-                         * Parameters to be used for each run of this job. The SQL alert task does not support custom parameters.
-                         */
-                        parameters?: {
-                            [k: string]: string;
-                        };
-                        /**
-                         * If query, indicates that this job must execute a SQL query.
-                         */
-                        query?: {
-                            /**
-                             * The canonical identifier of the SQL query.
-                             */
-                            query_id: string;
-                        };
-                        /**
-                         * The canonical identifier of the SQL warehouse. Only serverless and pro SQL warehouses are supported.
-                         */
-                        warehouse_id: string;
-                    };
-                    /**
-                     * A unique name for the task. This field is used to refer to this task from other tasks.
-                     * This field is required and must be unique within its parent job.
-                     * On Update or Reset, this field is used to reference the tasks to be updated or reset.
-                     * The maximum length is 100 characters.
-                     */
-                    task_key: string;
-                    /**
-                     * An optional timeout applied to each run of this job task. The default behavior is to have no timeout.
-                     */
-                    timeout_seconds?: number;
-                }[];
-                /**
-                 * An optional timeout applied to each run of this job. The default behavior is to have no timeout.
-                 */
-                timeout_seconds?: number;
-                /**
-                 * Trigger settings for the job. Can be used to trigger a run when new files arrive in an external location. The default behavior is that the job runs only when triggered by clicking “Run Now” in the Jobs UI or sending an API request to `runNow`.
-                 */
-                trigger?: {
-                    /**
-                     * File arrival trigger settings.
-                     */
-                    file_arrival?: {
-                        /**
-                         * If set, the trigger starts a run only after the specified amount of time passed since
-                         * the last time the trigger fired. The minimum allowed value is 60 seconds
-                         *
-                         */
-                        min_time_between_triggers_seconds?: number;
-                        /**
-                         * URL to be monitored for file arrivals. The path must point to the root or a subpath of the external location.
-                         */
-                        url?: string;
-                        /**
-                         * If set, the trigger starts a run only after no file activity has occurred for the specified amount of time.
-                         * This makes it possible to wait for a batch of incoming files to arrive before triggering a run. The
-                         * minimum allowed value is 60 seconds.
-                         *
-                         */
-                        wait_after_last_change_seconds?: number;
-                    };
-                    /**
-                     * Whether this trigger is paused or not.
-                     */
-                    pause_status?: string;
-                };
-                /**
-                 * A collection of system notification IDs to notify when the run begins or completes. The default behavior is to not send any system notifications.
-                 */
-                webhook_notifications?: {
-                    on_duration_warning_threshold_exceeded?: {
-                        id?: string;
-                    }[];
-                    /**
-                     * An optional list of system notification IDs to call when the run fails. A maximum of 3 destinations can be specified for the `on_failure` property.
-                     */
-                    on_failure?: {
-                        id?: string;
-                    }[];
-                    /**
-                     * An optional list of system notification IDs to call when the run starts. A maximum of 3 destinations can be specified for the `on_start` property.
-                     */
-                    on_start?: {
-                        id?: string;
-                    }[];
-                    /**
-                     * An optional list of system notification IDs to call when the run completes successfully. A maximum of 3 destinations can be specified for the `on_success` property.
-                     */
-                    on_success?: {
-                        id?: string;
-                    }[];
-                };
-            };
-        };
-        model_serving_endpoints?: {
-            [k: string]: {
-                config: {
-                    served_models: {
-                        environment_vars?: {
-                            [k: string]: string;
-                        };
-                        instance_profile_arn?: string;
-                        model_name: string;
-                        model_version: string;
-                        name?: string;
-                        scale_to_zero_enabled: boolean;
-                        workload_size: string;
-                    }[];
-                    traffic_config?: {
-                        routes?: {
-                            served_model_name: string;
-                            traffic_percentage: number;
-                        }[];
-                    };
-                };
-                name: string;
-                permissions?: {
-                    group_name?: string;
-                    level: string;
-                    service_principal_name?: string;
-                    user_name?: string;
-                }[];
-            };
-        };
-        /**
-         * List of MLflow models
-         */
-        models?: {
-            [k: string]: {
-                /**
-                 * Timestamp recorded when this `registered_model` was created.
-                 */
-                creation_timestamp?: number;
-                /**
-                 * Description of this `registered_model`.
-                 */
-                description?: string;
-                /**
-                 * Timestamp recorded when metadata for this `registered_model` was last updated.
-                 */
-                last_updated_timestamp?: number;
-                /**
-                 * Collection of latest model versions for each stage.
-                 * Only contains models with current `READY` status.
-                 */
-                latest_versions?: {
-                    /**
-                     * Timestamp recorded when this `model_version` was created.
-                     */
-                    creation_timestamp?: number;
-                    /**
-                     * Current stage for this `model_version`.
-                     */
-                    current_stage?: string;
-                    /**
-                     * Description of this `model_version`.
-                     */
-                    description?: string;
-                    /**
-                     * Timestamp recorded when metadata for this `model_version` was last updated.
-                     */
-                    last_updated_timestamp?: number;
-                    /**
-                     * Unique name of the model
-                     */
-                    name?: string;
-                    /**
-                     * MLflow run ID used when creating `model_version`, if `source` was generated by an
-                     * experiment run stored in MLflow tracking server.
-                     */
-                    run_id?: string;
-                    /**
-                     * Run Link: Direct link to the run that generated this version
-                     */
-                    run_link?: string;
-                    /**
-                     * URI indicating the location of the source model artifacts, used when creating `model_version`
-                     */
-                    source?: string;
-                    /**
-                     * Current status of `model_version`
-                     */
-                    status?: string;
-                    /**
-                     * Details on current `status`, if it is pending or failed.
-                     */
-                    status_message?: string;
-                    /**
-                     * Tags: Additional metadata key-value pairs for this `model_version`.
-                     */
-                    tags?: {
-                        /**
-                         * The tag key.
-                         */
-                        key?: string;
-                        /**
-                         * The tag value.
-                         */
-                        value?: string;
-                    }[];
-                    /**
-                     * User that created this `model_version`.
-                     */
-                    user_id?: string;
-                    /**
-                     * Model's version number.
-                     */
-                    version?: string;
-                }[];
-                /**
-                 * Unique name for the model.
-                 */
-                name?: string;
-                permissions?: {
-                    group_name?: string;
-                    level: string;
-                    service_principal_name?: string;
-                    user_name?: string;
-                }[];
-                /**
-                 * Tags: Additional metadata key-value pairs for this `registered_model`.
-                 */
-                tags?: {
-                    /**
-                     * The tag key.
-                     */
-                    key?: string;
-                    /**
-                     * The tag value.
-                     */
-                    value?: string;
-                }[];
-                /**
-                 * User that created this `registered_model`
-                 */
-                user_id?: string;
-            };
-        };
-        /**
-         * List of DLT pipelines
-         */
-        pipelines?: {
-            [k: string]: {
-                /**
-                 * A catalog in Unity Catalog to publish data from this pipeline to. If `target` is specified, tables in this pipeline are published to a `target` schema inside `catalog` (for example, `catalog`.`target`.`table`). If `target` is not specified, no data is published to Unity Catalog.
-                 */
-                catalog?: string;
-                /**
-                 * DLT Release Channel that specifies which version to use.
-                 */
-                channel?: string;
-                /**
-                 * Cluster settings for this pipeline deployment.
-                 */
-                clusters?: {
-                    /**
-                     * Note: This field won't be persisted. Only API users will check this field.
-                     */
-                    apply_policy_default_values?: boolean;
-                    /**
-                     * Parameters needed in order to automatically scale clusters up and down based on load.
-                     * Note: autoscaling works best with DB runtime versions 3.0 or later.
-                     */
-                    autoscale?: {
-                        /**
-                         * The maximum number of workers to which the cluster can scale up when overloaded.
-                         * Note that `max_workers` must be strictly greater than `min_workers`.
-                         */
-                        max_workers: number;
-                        /**
-                         * The minimum number of workers to which the cluster can scale down when underutilized.
-                         * It is also the initial number of workers the cluster will have after creation.
-                         */
-                        min_workers: number;
-                    };
-                    /**
-                     * Attributes related to clusters running on Amazon Web Services.
-                     * If not specified at cluster creation, a set of default values will be used.
-                     */
-                    aws_attributes?: {
-                        availability?: string;
-                        /**
-                         * The number of volumes launched for each instance. Users can choose up to 10 volumes.
-                         * This feature is only enabled for supported node types. Legacy node types cannot specify
-                         * custom EBS volumes.
-                         * For node types with no instance store, at least one EBS volume needs to be specified;
-                         * otherwise, cluster creation will fail.
-                         *
-                         * These EBS volumes will be mounted at `/ebs0`, `/ebs1`, and etc.
-                         * Instance store volumes will be mounted at `/local_disk0`, `/local_disk1`, and etc.
-                         *
-                         * If EBS volumes are attached, Databricks will configure Spark to use only the EBS volumes for
-                         * scratch storage because heterogenously sized scratch devices can lead to inefficient disk
-                         * utilization. If no EBS volumes are attached, Databricks will configure Spark to use instance
-                         * store volumes.
-                         *
-                         * Please note that if EBS volumes are specified, then the Spark configuration `spark.local.dir`
-                         * will be overridden.
-                         */
-                        ebs_volume_count?: number;
-                        /**
-                         * <needs content added>
-                         */
-                        ebs_volume_iops?: number;
-                        /**
-                         * The size of each EBS volume (in GiB) launched for each instance. For general purpose
-                         * SSD, this value must be within the range 100 - 4096. For throughput optimized HDD,
-                         * this value must be within the range 500 - 4096.
-                         */
-                        ebs_volume_size?: number;
-                        /**
-                         * <needs content added>
-                         */
-                        ebs_volume_throughput?: number;
-                        ebs_volume_type?: string;
-                        /**
-                         * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                         * If this value is greater than 0, the cluster driver node in particular will be placed on an
-                         * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                         * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                         * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                         * be placed on `availability` instances. Note that this value does not affect
-                         * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                         */
-                        first_on_demand?: number;
-                        /**
-                         * Nodes for this cluster will only be placed on AWS instances with this instance profile. If
-                         * ommitted, nodes will be placed on instances without an IAM instance profile. The instance
-                         * profile must have previously been added to the Databricks environment by an account
-                         * administrator.
-                         *
-                         * This feature may only be available to certain customer plans.
-                         *
-                         * If this field is ommitted, we will pull in the default from the conf if it exists.
-                         */
-                        instance_profile_arn?: string;
-                        /**
-                         * The bid price for AWS spot instances, as a percentage of the corresponding instance type's
-                         * on-demand price.
-                         * For example, if this field is set to 50, and the cluster needs a new `r3.xlarge` spot
-                         * instance, then the bid price is half of the price of
-                         * on-demand `r3.xlarge` instances. Similarly, if this field is set to 200, the bid price is twice
-                         * the price of on-demand `r3.xlarge` instances. If not specified, the default value is 100.
-                         * When spot instances are requested for this cluster, only spot instances whose bid price
-                         * percentage matches this field will be considered.
-                         * Note that, for safety, we enforce this field to be no more than 10000.
-                         *
-                         * The default value and documentation here should be kept consistent with
-                         * CommonConf.defaultSpotBidPricePercent and CommonConf.maxSpotBidPricePercent.
-                         */
-                        spot_bid_price_percent?: number;
-                        /**
-                         * Identifier for the availability zone/datacenter in which the cluster resides.
-                         * This string will be of a form like "us-west-2a". The provided availability
-                         * zone must be in the same region as the Databricks deployment. For example, "us-west-2a"
-                         * is not a valid zone id if the Databricks deployment resides in the "us-east-1" region.
-                         * This is an optional field at cluster creation, and if not specified, a default zone will be used.
-                         * If the zone specified is "auto", will try to place cluster in a zone with high availability,
-                         * and will retry placement in a different AZ if there is not enough capacity.
-                         * See [[AutoAZHelper.scala]] for more details.
-                         * The list of available zones as well as the default value can be found by using the
-                         * `List Zones`_ method.
-                         */
-                        zone_id?: string;
-                    };
-                    /**
-                     * Attributes related to clusters running on Microsoft Azure.
-                     * If not specified at cluster creation, a set of default values will be used.
-                     */
-                    azure_attributes?: {
-                        availability?: string;
-                        /**
-                         * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                         * This value should be greater than 0, to make sure the cluster driver node is placed on an
-                         * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                         * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                         * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                         * be placed on `availability` instances. Note that this value does not affect
-                         * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                         */
-                        first_on_demand?: number;
-                        /**
-                         * Defines values necessary to configure and run Azure Log Analytics agent
-                         */
-                        log_analytics_info?: {
-                            /**
-                             * <needs content added>
-                             */
-                            log_analytics_primary_key?: string;
-                            /**
-                             * <needs content added>
-                             */
-                            log_analytics_workspace_id?: string;
-                        };
-                        /**
-                         * The max bid price to be used for Azure spot instances.
-                         * The Max price for the bid cannot be higher than the on-demand price of the instance.
-                         * If not specified, the default value is -1, which specifies that the instance cannot be evicted
-                         * on the basis of price, and only on the basis of availability. Further, the value should > 0 or -1.
-                         */
-                        spot_bid_max_price?: number;
-                    };
-                    /**
-                     * The configuration for delivering spark logs to a long-term storage destination.
-                     * Only dbfs destinations are supported. Only one destination can be specified
-                     * for one cluster. If the conf is given, the logs will be delivered to the destination every
-                     * `5 mins`. The destination of driver logs is `$destination/$clusterId/driver`, while
-                     * the destination of executor logs is `$destination/$clusterId/executor`.
-                     *
-                     */
-                    cluster_log_conf?: {
-                        /**
-                         * destination needs to be provided. e.g.
-                         * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                         */
-                        dbfs?: {
-                            /**
-                             * dbfs destination, e.g. `dbfs:/my/path`
-                             */
-                            destination?: string;
-                        };
-                        /**
-                         * destination and either the region or endpoint need to be provided. e.g.
-                         * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                         * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                         * `instance_profile_arn` has permission to write data to the s3 destination.
-                         */
-                        s3?: {
-                            /**
-                             * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                             * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                             * the destination bucket and prefix. The full list of possible canned acl can be found at
-                             * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                             * Please also note that by default only the object owner gets full controls. If you are using cross account
-                             * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                             * read the logs.
-                             */
-                            canned_acl?: string;
-                            /**
-                             * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                             * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                             * destination. Please also note that you cannot use AWS keys to deliver logs.
-                             */
-                            destination?: string;
-                            /**
-                             * (Optional) Flag to enable server side encryption, `false` by default.
-                             */
-                            enable_encryption?: boolean;
-                            /**
-                             * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                             * encryption is enabled and the default type is `sse-s3`.
-                             */
-                            encryption_type?: string;
-                            /**
-                             * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                             * If both are set, endpoint will be used.
-                             */
-                            endpoint?: string;
-                            /**
-                             * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                             */
-                            kms_key?: string;
-                            /**
-                             * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                             * endpoint will be used.
-                             */
-                            region?: string;
-                        };
-                    };
-                    /**
-                     * Additional tags for cluster resources. Databricks will tag all cluster resources (e.g., AWS
-                     * instances and EBS volumes) with these tags in addition to `default_tags`. Notes:
-                     *
-                     * - Currently, Databricks allows at most 45 custom tags
-                     *
-                     * - Clusters can only reuse cloud resources if the resources' tags are a subset of the cluster tags
-                     */
-                    custom_tags?: {
-                        [k: string]: string;
-                    };
-                    /**
-                     * The optional ID of the instance pool for the driver of the cluster belongs.
-                     * The pool cluster uses the instance pool with id (instance_pool_id) if the driver pool is not
-                     * assigned.
-                     */
-                    driver_instance_pool_id?: string;
-                    /**
-                     * The node type of the Spark driver.
-                     * Note that this field is optional; if unset, the driver node type will be set as the same value
-                     * as `node_type_id` defined above.
-                     */
-                    driver_node_type_id?: string;
-                    /**
-                     * Attributes related to clusters running on Google Cloud Platform.
-                     * If not specified at cluster creation, a set of default values will be used.
-                     */
-                    gcp_attributes?: {
-                        availability?: string;
-                        /**
-                         * boot disk size in GB
-                         */
-                        boot_disk_size?: number;
-                        /**
-                         * If provided, the cluster will impersonate the google service account when accessing
-                         * gcloud services (like GCS). The google service account
-                         * must have previously been added to the Databricks environment by an account
-                         * administrator.
-                         */
-                        google_service_account?: string;
-                        /**
-                         * If provided, each node (workers and driver) in the cluster will have this number of local SSDs attached. Each local SSD is 375GB in size. Refer to [GCP documentation](https://cloud.google.com/compute/docs/disks/local-ssd#choose_number_local_ssds) for the supported number of local SSDs for each instance type.
-                         */
-                        local_ssd_count?: number;
-                    };
-                    /**
-                     * The optional ID of the instance pool to which the cluster belongs.
-                     */
-                    instance_pool_id?: string;
-                    /**
-                     * A label for the cluster specification, either `default` to configure the default cluster, or `maintenance` to configure the maintenance cluster. This field is optional. The default value is `default`.
-                     */
-                    label?: string;
-                    /**
-                     * This field encodes, through a single value, the resources available to each of
-                     * the Spark nodes in this cluster. For example, the Spark nodes can be provisioned
-                     * and optimized for memory or compute intensive workloads. A list of available node
-                     * types can be retrieved by using the :method:clusters/listNodeTypes API call.
-                     *
-                     */
-                    node_type_id?: string;
-                    /**
-                     * Number of worker nodes that this cluster should have. A cluster has one Spark Driver
-                     * and `num_workers` Executors for a total of `num_workers` + 1 Spark nodes.
-                     *
-                     * Note: When reading the properties of a cluster, this field reflects the desired number
-                     * of workers rather than the actual current number of workers. For instance, if a cluster
-                     * is resized from 5 to 10 workers, this field will immediately be updated to reflect
-                     * the target size of 10 workers, whereas the workers listed in `spark_info` will gradually
-                     * increase from 5 to 10 as the new nodes are provisioned.
-                     */
-                    num_workers?: number;
-                    /**
-                     * The ID of the cluster policy used to create the cluster if applicable.
-                     */
-                    policy_id?: string;
-                    /**
-                     * An object containing a set of optional, user-specified Spark configuration key-value pairs.
-                     * See :method:clusters/create for more details.
-                     *
-                     */
-                    spark_conf?: {
-                        [k: string]: string;
-                    };
-                    /**
-                     * An object containing a set of optional, user-specified environment variable key-value pairs.
-                     * Please note that key-value pair of the form (X,Y) will be exported as is (i.e.,
-                     * `export X='Y'`) while launching the driver and workers.
-                     *
-                     * In order to specify an additional set of `SPARK_DAEMON_JAVA_OPTS`, we recommend appending
-                     * them to `$SPARK_DAEMON_JAVA_OPTS` as shown in the example below. This ensures that all
-                     * default databricks managed environmental variables are included as well.
-                     *
-                     * Example Spark environment variables:
-                     * `{"SPARK_WORKER_MEMORY": "28000m", "SPARK_LOCAL_DIRS": "/local_disk0"}` or
-                     * `{"SPARK_DAEMON_JAVA_OPTS": "$SPARK_DAEMON_JAVA_OPTS -Dspark.shuffle.service.enabled=true"}`
-                     */
-                    spark_env_vars?: {
-                        [k: string]: string;
-                    };
-                    /**
-                     * SSH public key contents that will be added to each Spark node in this cluster. The
-                     * corresponding private keys can be used to login with the user name `ubuntu` on port `2200`.
-                     * Up to 10 keys can be specified.
-                     */
-                    ssh_public_keys?: string[];
-                }[];
-                /**
-                 * String-String configuration for this pipeline execution.
-                 */
-                configuration?: {
-                    [k: string]: string;
-                };
-                /**
-                 * Whether the pipeline is continuous or triggered. This replaces `trigger`.
-                 */
-                continuous?: boolean;
-                /**
-                 * Whether the pipeline is in Development mode. Defaults to false.
-                 */
-                development?: boolean;
-                /**
-                 * Pipeline product edition.
-                 */
-                edition?: string;
-                /**
-                 * Filters on which Pipeline packages to include in the deployed graph.
-                 */
-                filters?: {
-                    /**
-                     * Paths to exclude.
-                     */
-                    exclude?: string[];
-                    /**
-                     * Paths to include.
-                     */
-                    include?: string[];
-                };
-                /**
-                 * Unique identifier for this pipeline.
-                 */
-                id?: string;
-                /**
-                 * Libraries or code needed by this deployment.
-                 */
-                libraries?: {
-                    /**
-                     * The path to a file that defines a pipeline and is stored in the Databricks Repos.
-                     *
-                     */
-                    file?: {
-                        /**
-                         * The absolute path of the file.
-                         */
-                        path?: string;
-                    };
-                    /**
-                     * URI of the jar to be installed. Currently only DBFS is supported.
-                     *
-                     */
-                    jar?: string;
-                    /**
-                     * Specification of a maven library to be installed.
-                     *
-                     */
-                    maven?: {
-                        /**
-                         * Gradle-style maven coordinates. For example: "org.jsoup:jsoup:1.7.2".
-                         */
-                        coordinates: string;
-                        /**
-                         * List of dependences to exclude. For example: `["slf4j:slf4j", "*:hadoop-client"]`.
-                         *
-                         * Maven dependency exclusions:
-                         * https://maven.apache.org/guides/introduction/introduction-to-optional-and-excludes-dependencies.html.
-                         */
-                        exclusions?: string[];
-                        /**
-                         * Maven repo to install the Maven package from. If omitted, both Maven Central Repository
-                         * and Spark Packages are searched.
-                         */
-                        repo?: string;
-                    };
-                    /**
-                     * The path to a notebook that defines a pipeline and is stored in the <Databricks> workspace.
-                     *
-                     */
-                    notebook?: {
-                        /**
-                         * The absolute path of the notebook.
-                         */
-                        path?: string;
-                    };
-                }[];
-                /**
-                 * Friendly identifier for this pipeline.
-                 */
-                name?: string;
-                permissions?: {
-                    group_name?: string;
-                    level: string;
-                    service_principal_name?: string;
-                    user_name?: string;
-                }[];
-                /**
-                 * Whether Photon is enabled for this pipeline.
-                 */
-                photon?: boolean;
-                /**
-                 * Whether serverless compute is enabled for this pipeline.
-                 */
-                serverless?: boolean;
-                /**
-                 * DBFS root directory for storing checkpoints and tables.
-                 */
-                storage?: string;
-                /**
-                 * Target schema (database) to add tables in this pipeline to. If not specified, no data is published to the Hive metastore or Unity Catalog. To publish to Unity Catalog, also specify `catalog`.
-                 */
-                target?: string;
-                /**
-                 * Which pipeline trigger to use. Deprecated: Use `continuous` instead.
-                 */
-                trigger?: {
-                    cron?: {
-                        quartz_cron_schedule?: string;
-                        timezone_id?: string;
-                    };
-                    manual?: unknown;
-                };
-            };
-        };
-    };
-    run_as?: {
-        service_principal_name?: string;
-        user_name?: string;
-    };
-    sync?: {
-        exclude?: string[];
-        include?: string[];
-    };
-    targets?: {
-        [k: string]: {
-            /**
-             * A description of all code artifacts in this bundle.
-             */
-            artifacts?: {
-                [k: string]: {
-                    build: string;
-                    files: {
-                        source: string;
-                    }[];
-                    path: string;
-                    type: string;
-                };
-            };
-            /**
-             * The details for this bundle.
-             */
-            bundle?: {
-                compute_id?: string;
-                git?: {
-                    branch?: string;
-                    origin_url?: string;
-                };
-                /**
-                 * The name of the bundle.
-                 */
-                name: string;
-            };
-            compute_id?: string;
-            default?: boolean;
-            git?: {
-                branch?: string;
-                origin_url?: string;
-            };
-            mode?: string;
-            /**
-             * Collection of Databricks resources to deploy.
-             */
-            resources?: {
-                /**
-                 * List of MLflow experiments
-                 */
-                experiments?: {
-                    [k: string]: {
-                        /**
-                         * Location where artifacts for the experiment are stored.
-                         */
-                        artifact_location?: string;
-                        /**
-                         * Creation time
-                         */
-                        creation_time?: number;
-                        /**
-                         * Unique identifier for the experiment.
-                         */
-                        experiment_id?: string;
-                        /**
-                         * Last update time
-                         */
-                        last_update_time?: number;
-                        /**
-                         * Current life cycle stage of the experiment: "active" or "deleted".
-                         * Deleted experiments are not returned by APIs.
-                         */
-                        lifecycle_stage?: string;
-                        /**
-                         * Human readable name that identifies the experiment.
-                         */
-                        name?: string;
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                        /**
-                         * Tags: Additional metadata key-value pairs.
-                         */
-                        tags?: {
-                            /**
-                             * The tag key.
-                             */
-                            key?: string;
-                            /**
-                             * The tag value.
-                             */
-                            value?: string;
-                        }[];
-                    };
-                };
-                /**
-                 * List of Databricks jobs
-                 */
-                jobs?: {
-                    [k: string]: {
-                        /**
-                         * A list of compute requirements that can be referenced by tasks of this job.
-                         */
-                        compute?: {
-                            /**
-                             * A unique name for the compute requirement. This field is required and must be unique within the job.
-                             * `JobTaskSettings` may refer to this field to determine the compute requirements for the task execution.
-                             */
-                            compute_key: string;
-                            spec: {
-                                /**
-                                 * The kind of compute described by this compute specification.
-                                 */
-                                kind?: string;
-                            };
-                        }[];
-                        /**
-                         * An optional continuous property for this job. The continuous property will ensure that there is always one run executing. Only one of `schedule` and `continuous` can be used.
-                         */
-                        continuous?: {
-                            /**
-                             * Whether this trigger is paused or not.
-                             */
-                            pause_status?: string;
-                        };
-                        /**
-                         * An optional set of email addresses that is notified when runs of this job begin or complete as well as when this job is deleted. The default behavior is to not send any emails.
-                         */
-                        email_notifications?: {
-                            /**
-                             * If true, do not send email to recipients specified in `on_failure` if the run is skipped.
-                             */
-                            no_alert_for_skipped_runs?: boolean;
-                            on_duration_warning_threshold_exceeded?: string[];
-                            /**
-                             * A list of email addresses to be notified when a run unsuccessfully completes. A run is considered to have completed unsuccessfully if it ends with an `INTERNAL_ERROR` `life_cycle_state` or a `FAILED`, or `TIMED_OUT` result_state. If this is not specified on job creation, reset, or update the list is empty, and notifications are not sent.
-                             */
-                            on_failure?: string[];
-                            /**
-                             * A list of email addresses to be notified when a run begins. If not specified on job creation, reset, or update, the list is empty, and notifications are not sent.
-                             */
-                            on_start?: string[];
-                            /**
-                             * A list of email addresses to be notified when a run successfully completes. A run is considered to have completed successfully if it ends with a `TERMINATED` `life_cycle_state` and a `SUCCESS` result_state. If not specified on job creation, reset, or update, the list is empty, and notifications are not sent.
-                             */
-                            on_success?: string[];
-                        };
-                        /**
-                         * Used to tell what is the format of the job. This field is ignored in Create/Update/Reset calls. When using the Jobs API 2.1 this value is always set to `"MULTI_TASK"`.
-                         */
-                        format?: string;
-                        /**
-                         * An optional specification for a remote repository containing the notebooks used by this job's notebook tasks.
-                         */
-                        git_source?: {
-                            /**
-                             * Name of the branch to be checked out and used by this job.
-                             * This field cannot be specified in conjunction with git_tag or git_commit.
-                             *
-                             * The maximum length is 255 characters.
-                             *
-                             */
-                            git_branch?: string;
-                            /**
-                             * Commit to be checked out and used by this job. This field cannot be specified in conjunction with git_branch or git_tag.
-                             * The maximum length is 64 characters.
-                             */
-                            git_commit?: string;
-                            /**
-                             * Unique identifier of the service used to host the Git repository. The value is case insensitive.
-                             */
-                            git_provider: string;
-                            git_snapshot?: {
-                                /**
-                                 * Commit that was used to execute the run. If git_branch was specified, this points to the HEAD of the branch at the time of the run; if git_tag was specified, this points to the commit the tag points to.
-                                 */
-                                used_commit?: string;
-                            };
-                            /**
-                             * Name of the tag to be checked out and used by this job.
-                             * This field cannot be specified in conjunction with git_branch or git_commit.
-                             *
-                             * The maximum length is 255 characters.
-                             *
-                             */
-                            git_tag?: string;
-                            /**
-                             * URL of the repository to be cloned by this job.
-                             * The maximum length is 300 characters.
-                             */
-                            git_url: string;
-                            job_source?: {
-                                dirty_state?: string;
-                                import_from_git_branch: string;
-                                job_config_path: string;
-                            };
-                        };
-                        health?: {
-                            rules?: {
-                                metric?: string;
-                                op?: string;
-                                value?: number;
-                            }[];
-                        };
-                        /**
-                         * A list of job cluster specifications that can be shared and reused by tasks of this job. Libraries cannot be declared in a shared job cluster. You must declare dependent libraries in task settings.
-                         */
-                        job_clusters?: {
-                            /**
-                             * A unique name for the job cluster. This field is required and must be unique within the job.
-                             * `JobTaskSettings` may refer to this field to determine which cluster to launch for the task execution.
-                             */
-                            job_cluster_key: string;
-                            /**
-                             * If new_cluster, a description of a cluster that is created for only for this task.
-                             */
-                            new_cluster?: {
-                                /**
-                                 * Parameters needed in order to automatically scale clusters up and down based on load.
-                                 * Note: autoscaling works best with DB runtime versions 3.0 or later.
-                                 */
-                                autoscale?: {
-                                    /**
-                                     * The maximum number of workers to which the cluster can scale up when overloaded.
-                                     * Note that `max_workers` must be strictly greater than `min_workers`.
-                                     */
-                                    max_workers: number;
-                                    /**
-                                     * The minimum number of workers to which the cluster can scale down when underutilized.
-                                     * It is also the initial number of workers the cluster will have after creation.
-                                     */
-                                    min_workers: number;
-                                };
-                                /**
-                                 * Automatically terminates the cluster after it is inactive for this time in minutes. If not set,
-                                 * this cluster will not be automatically terminated. If specified, the threshold must be between
-                                 * 10 and 10000 minutes.
-                                 * Users can also set this value to 0 to explicitly disable automatic termination.
-                                 */
-                                autotermination_minutes?: number;
-                                /**
-                                 * Attributes related to clusters running on Amazon Web Services.
-                                 * If not specified at cluster creation, a set of default values will be used.
-                                 */
-                                aws_attributes?: {
-                                    availability?: string;
-                                    /**
-                                     * The number of volumes launched for each instance. Users can choose up to 10 volumes.
-                                     * This feature is only enabled for supported node types. Legacy node types cannot specify
-                                     * custom EBS volumes.
-                                     * For node types with no instance store, at least one EBS volume needs to be specified;
-                                     * otherwise, cluster creation will fail.
-                                     *
-                                     * These EBS volumes will be mounted at `/ebs0`, `/ebs1`, and etc.
-                                     * Instance store volumes will be mounted at `/local_disk0`, `/local_disk1`, and etc.
-                                     *
-                                     * If EBS volumes are attached, Databricks will configure Spark to use only the EBS volumes for
-                                     * scratch storage because heterogenously sized scratch devices can lead to inefficient disk
-                                     * utilization. If no EBS volumes are attached, Databricks will configure Spark to use instance
-                                     * store volumes.
-                                     *
-                                     * Please note that if EBS volumes are specified, then the Spark configuration `spark.local.dir`
-                                     * will be overridden.
-                                     */
-                                    ebs_volume_count?: number;
-                                    /**
-                                     * <needs content added>
-                                     */
-                                    ebs_volume_iops?: number;
-                                    /**
-                                     * The size of each EBS volume (in GiB) launched for each instance. For general purpose
-                                     * SSD, this value must be within the range 100 - 4096. For throughput optimized HDD,
-                                     * this value must be within the range 500 - 4096.
-                                     */
-                                    ebs_volume_size?: number;
-                                    /**
-                                     * <needs content added>
-                                     */
-                                    ebs_volume_throughput?: number;
-                                    ebs_volume_type?: string;
-                                    /**
-                                     * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                                     * If this value is greater than 0, the cluster driver node in particular will be placed on an
-                                     * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                                     * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                                     * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                                     * be placed on `availability` instances. Note that this value does not affect
-                                     * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                                     */
-                                    first_on_demand?: number;
-                                    /**
-                                     * Nodes for this cluster will only be placed on AWS instances with this instance profile. If
-                                     * ommitted, nodes will be placed on instances without an IAM instance profile. The instance
-                                     * profile must have previously been added to the Databricks environment by an account
-                                     * administrator.
-                                     *
-                                     * This feature may only be available to certain customer plans.
-                                     *
-                                     * If this field is ommitted, we will pull in the default from the conf if it exists.
-                                     */
-                                    instance_profile_arn?: string;
-                                    /**
-                                     * The bid price for AWS spot instances, as a percentage of the corresponding instance type's
-                                     * on-demand price.
-                                     * For example, if this field is set to 50, and the cluster needs a new `r3.xlarge` spot
-                                     * instance, then the bid price is half of the price of
-                                     * on-demand `r3.xlarge` instances. Similarly, if this field is set to 200, the bid price is twice
-                                     * the price of on-demand `r3.xlarge` instances. If not specified, the default value is 100.
-                                     * When spot instances are requested for this cluster, only spot instances whose bid price
-                                     * percentage matches this field will be considered.
-                                     * Note that, for safety, we enforce this field to be no more than 10000.
-                                     *
-                                     * The default value and documentation here should be kept consistent with
-                                     * CommonConf.defaultSpotBidPricePercent and CommonConf.maxSpotBidPricePercent.
-                                     */
-                                    spot_bid_price_percent?: number;
-                                    /**
-                                     * Identifier for the availability zone/datacenter in which the cluster resides.
-                                     * This string will be of a form like "us-west-2a". The provided availability
-                                     * zone must be in the same region as the Databricks deployment. For example, "us-west-2a"
-                                     * is not a valid zone id if the Databricks deployment resides in the "us-east-1" region.
-                                     * This is an optional field at cluster creation, and if not specified, a default zone will be used.
-                                     * If the zone specified is "auto", will try to place cluster in a zone with high availability,
-                                     * and will retry placement in a different AZ if there is not enough capacity.
-                                     * See [[AutoAZHelper.scala]] for more details.
-                                     * The list of available zones as well as the default value can be found by using the
-                                     * `List Zones`_ method.
-                                     */
-                                    zone_id?: string;
-                                };
-                                /**
-                                 * Attributes related to clusters running on Microsoft Azure.
-                                 * If not specified at cluster creation, a set of default values will be used.
-                                 */
-                                azure_attributes?: {
-                                    availability?: string;
-                                    /**
-                                     * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                                     * This value should be greater than 0, to make sure the cluster driver node is placed on an
-                                     * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                                     * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                                     * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                                     * be placed on `availability` instances. Note that this value does not affect
-                                     * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                                     */
-                                    first_on_demand?: number;
-                                    /**
-                                     * Defines values necessary to configure and run Azure Log Analytics agent
-                                     */
-                                    log_analytics_info?: {
-                                        /**
-                                         * <needs content added>
-                                         */
-                                        log_analytics_primary_key?: string;
-                                        /**
-                                         * <needs content added>
-                                         */
-                                        log_analytics_workspace_id?: string;
-                                    };
-                                    /**
-                                     * The max bid price to be used for Azure spot instances.
-                                     * The Max price for the bid cannot be higher than the on-demand price of the instance.
-                                     * If not specified, the default value is -1, which specifies that the instance cannot be evicted
-                                     * on the basis of price, and only on the basis of availability. Further, the value should > 0 or -1.
-                                     */
-                                    spot_bid_max_price?: number;
-                                };
-                                /**
-                                 * The configuration for delivering spark logs to a long-term storage destination.
-                                 * Two kinds of destinations (dbfs and s3) are supported. Only one destination can be specified
-                                 * for one cluster. If the conf is given, the logs will be delivered to the destination every
-                                 * `5 mins`. The destination of driver logs is `$destination/$clusterId/driver`, while
-                                 * the destination of executor logs is `$destination/$clusterId/executor`.
-                                 */
-                                cluster_log_conf?: {
-                                    /**
-                                     * destination needs to be provided. e.g.
-                                     * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                                     */
-                                    dbfs?: {
-                                        /**
-                                         * dbfs destination, e.g. `dbfs:/my/path`
-                                         */
-                                        destination?: string;
-                                    };
-                                    /**
-                                     * destination and either the region or endpoint need to be provided. e.g.
-                                     * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                                     * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                                     * `instance_profile_arn` has permission to write data to the s3 destination.
-                                     */
-                                    s3?: {
-                                        /**
-                                         * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                                         * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                                         * the destination bucket and prefix. The full list of possible canned acl can be found at
-                                         * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                                         * Please also note that by default only the object owner gets full controls. If you are using cross account
-                                         * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                                         * read the logs.
-                                         */
-                                        canned_acl?: string;
-                                        /**
-                                         * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                                         * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                                         * destination. Please also note that you cannot use AWS keys to deliver logs.
-                                         */
-                                        destination?: string;
-                                        /**
-                                         * (Optional) Flag to enable server side encryption, `false` by default.
-                                         */
-                                        enable_encryption?: boolean;
-                                        /**
-                                         * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                                         * encryption is enabled and the default type is `sse-s3`.
-                                         */
-                                        encryption_type?: string;
-                                        /**
-                                         * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                                         * If both are set, endpoint will be used.
-                                         */
-                                        endpoint?: string;
-                                        /**
-                                         * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                                         */
-                                        kms_key?: string;
-                                        /**
-                                         * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                                         * endpoint will be used.
-                                         */
-                                        region?: string;
-                                    };
-                                };
-                                /**
-                                 * Cluster name requested by the user. This doesn't have to be unique.
-                                 * If not specified at creation, the cluster name will be an empty string.
-                                 *
-                                 */
-                                cluster_name?: string;
-                                cluster_source?: string;
-                                /**
-                                 * Additional tags for cluster resources. Databricks will tag all cluster resources (e.g., AWS
-                                 * instances and EBS volumes) with these tags in addition to `default_tags`. Notes:
-                                 *
-                                 * - Currently, Databricks allows at most 45 custom tags
-                                 *
-                                 * - Clusters can only reuse cloud resources if the resources' tags are a subset of the cluster tags
-                                 */
-                                custom_tags?: {
-                                    [k: string]: string;
-                                };
-                                data_security_mode?: string;
-                                docker_image?: {
-                                    basic_auth?: {
-                                        /**
-                                         * Password of the user
-                                         */
-                                        password?: string;
-                                        /**
-                                         * Name of the user
-                                         */
-                                        username?: string;
-                                    };
-                                    /**
-                                     * URL of the docker image.
-                                     */
-                                    url?: string;
-                                };
-                                /**
-                                 * The optional ID of the instance pool for the driver of the cluster belongs.
-                                 * The pool cluster uses the instance pool with id (instance_pool_id) if the driver pool is not
-                                 * assigned.
-                                 */
-                                driver_instance_pool_id?: string;
-                                /**
-                                 * The node type of the Spark driver. Note that this field is optional;
-                                 * if unset, the driver node type will be set as the same value
-                                 * as `node_type_id` defined above.
-                                 *
-                                 */
-                                driver_node_type_id?: string;
-                                /**
-                                 * Autoscaling Local Storage: when enabled, this cluster will dynamically acquire additional disk
-                                 * space when its Spark workers are running low on disk space. This feature requires specific AWS
-                                 * permissions to function correctly - refer to the User Guide for more details.
-                                 */
-                                enable_elastic_disk?: boolean;
-                                /**
-                                 * Whether to enable LUKS on cluster VMs' local disks
-                                 */
-                                enable_local_disk_encryption?: boolean;
-                                /**
-                                 * Attributes related to clusters running on Google Cloud Platform.
-                                 * If not specified at cluster creation, a set of default values will be used.
-                                 */
-                                gcp_attributes?: {
-                                    availability?: string;
-                                    /**
-                                     * boot disk size in GB
-                                     */
-                                    boot_disk_size?: number;
-                                    /**
-                                     * If provided, the cluster will impersonate the google service account when accessing
-                                     * gcloud services (like GCS). The google service account
-                                     * must have previously been added to the Databricks environment by an account
-                                     * administrator.
-                                     */
-                                    google_service_account?: string;
-                                    /**
-                                     * If provided, each node (workers and driver) in the cluster will have this number of local SSDs attached. Each local SSD is 375GB in size. Refer to [GCP documentation](https://cloud.google.com/compute/docs/disks/local-ssd#choose_number_local_ssds) for the supported number of local SSDs for each instance type.
-                                     */
-                                    local_ssd_count?: number;
-                                };
-                                /**
-                                 * The configuration for storing init scripts. Any number of destinations can be specified. The scripts are executed sequentially in the order provided. If `cluster_log_conf` is specified, init script logs are sent to `<destination>/<cluster-ID>/init_scripts`.
-                                 */
-                                init_scripts?: {
-                                    /**
-                                     * destination needs to be provided. e.g.
-                                     * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                                     */
-                                    dbfs?: {
-                                        /**
-                                         * dbfs destination, e.g. `dbfs:/my/path`
-                                         */
-                                        destination?: string;
-                                    };
-                                    /**
-                                     * destination and either the region or endpoint need to be provided. e.g.
-                                     * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                                     * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                                     * `instance_profile_arn` has permission to write data to the s3 destination.
-                                     */
-                                    s3?: {
-                                        /**
-                                         * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                                         * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                                         * the destination bucket and prefix. The full list of possible canned acl can be found at
-                                         * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                                         * Please also note that by default only the object owner gets full controls. If you are using cross account
-                                         * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                                         * read the logs.
-                                         */
-                                        canned_acl?: string;
-                                        /**
-                                         * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                                         * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                                         * destination. Please also note that you cannot use AWS keys to deliver logs.
-                                         */
-                                        destination?: string;
-                                        /**
-                                         * (Optional) Flag to enable server side encryption, `false` by default.
-                                         */
-                                        enable_encryption?: boolean;
-                                        /**
-                                         * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                                         * encryption is enabled and the default type is `sse-s3`.
-                                         */
-                                        encryption_type?: string;
-                                        /**
-                                         * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                                         * If both are set, endpoint will be used.
-                                         */
-                                        endpoint?: string;
-                                        /**
-                                         * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                                         */
-                                        kms_key?: string;
-                                        /**
-                                         * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                                         * endpoint will be used.
-                                         */
-                                        region?: string;
-                                    };
-                                    volumes?: {
-                                        destination?: string;
-                                    };
-                                    /**
-                                     * destination needs to be provided. e.g.
-                                     * `{ "workspace" : { "destination" : "/Users/user1@databricks.com/my-init.sh" } }`
-                                     */
-                                    workspace?: {
-                                        /**
-                                         * workspace files destination, e.g. `/Users/user1@databricks.com/my-init.sh`
-                                         */
-                                        destination?: string;
-                                    };
-                                }[];
-                                /**
-                                 * The optional ID of the instance pool to which the cluster belongs.
-                                 */
-                                instance_pool_id?: string;
-                                /**
-                                 * This field encodes, through a single value, the resources available to each of
-                                 * the Spark nodes in this cluster. For example, the Spark nodes can be provisioned
-                                 * and optimized for memory or compute intensive workloads. A list of available node
-                                 * types can be retrieved by using the :method:clusters/listNodeTypes API call.
-                                 *
-                                 */
-                                node_type_id?: string;
-                                /**
-                                 * Number of worker nodes that this cluster should have. A cluster has one Spark Driver
-                                 * and `num_workers` Executors for a total of `num_workers` + 1 Spark nodes.
-                                 *
-                                 * Note: When reading the properties of a cluster, this field reflects the desired number
-                                 * of workers rather than the actual current number of workers. For instance, if a cluster
-                                 * is resized from 5 to 10 workers, this field will immediately be updated to reflect
-                                 * the target size of 10 workers, whereas the workers listed in `spark_info` will gradually
-                                 * increase from 5 to 10 as the new nodes are provisioned.
-                                 */
-                                num_workers?: number;
-                                /**
-                                 * The ID of the cluster policy used to create the cluster if applicable.
-                                 */
-                                policy_id?: string;
-                                runtime_engine?: string;
-                                /**
-                                 * Single user name if data_security_mode is `SINGLE_USER`
-                                 */
-                                single_user_name?: string;
-                                /**
-                                 * An object containing a set of optional, user-specified Spark configuration key-value pairs.
-                                 * Users can also pass in a string of extra JVM options to the driver and the executors via
-                                 * `spark.driver.extraJavaOptions` and `spark.executor.extraJavaOptions` respectively.
-                                 *
-                                 */
-                                spark_conf?: {
-                                    [k: string]: string;
-                                };
-                                /**
-                                 * An object containing a set of optional, user-specified environment variable key-value pairs.
-                                 * Please note that key-value pair of the form (X,Y) will be exported as is (i.e.,
-                                 * `export X='Y'`) while launching the driver and workers.
-                                 *
-                                 * In order to specify an additional set of `SPARK_DAEMON_JAVA_OPTS`, we recommend appending
-                                 * them to `$SPARK_DAEMON_JAVA_OPTS` as shown in the example below. This ensures that all
-                                 * default databricks managed environmental variables are included as well.
-                                 *
-                                 * Example Spark environment variables:
-                                 * `{"SPARK_WORKER_MEMORY": "28000m", "SPARK_LOCAL_DIRS": "/local_disk0"}` or
-                                 * `{"SPARK_DAEMON_JAVA_OPTS": "$SPARK_DAEMON_JAVA_OPTS -Dspark.shuffle.service.enabled=true"}`
-                                 */
-                                spark_env_vars?: {
-                                    [k: string]: string;
-                                };
-                                /**
-                                 * The Spark version of the cluster, e.g. `3.3.x-scala2.11`.
-                                 * A list of available Spark versions can be retrieved by using
-                                 * the :method:clusters/sparkVersions API call.
-                                 *
-                                 */
-                                spark_version?: string;
-                                /**
-                                 * SSH public key contents that will be added to each Spark node in this cluster. The
-                                 * corresponding private keys can be used to login with the user name `ubuntu` on port `2200`.
-                                 * Up to 10 keys can be specified.
-                                 */
-                                ssh_public_keys?: string[];
-                                workload_type?: {
-                                    /**
-                                     *  defined what type of clients can use the cluster. E.g. Notebooks, Jobs
-                                     */
-                                    clients?: {
-                                        /**
-                                         * With jobs set, the cluster can be used for jobs
-                                         */
-                                        jobs?: boolean;
-                                        /**
-                                         * With notebooks set, this cluster can be used for notebooks
-                                         */
-                                        notebooks?: boolean;
-                                    };
-                                };
-                            };
-                        }[];
-                        /**
-                         * An optional maximum allowed number of concurrent runs of the job.
-                         *
-                         * Set this value if you want to be able to execute multiple runs of the same job concurrently. This is useful for example if you trigger your job on a frequent schedule and want to allow consecutive runs to overlap with each other, or if you want to trigger multiple runs which differ by their input parameters.
-                         *
-                         * This setting affects only new runs. For example, suppose the job’s concurrency is 4 and there are 4 concurrent active runs. Then setting the concurrency to 3 won’t kill any of the active runs. However, from then on, new runs are skipped unless there are fewer than 3 active runs.
-                         *
-                         * This value cannot exceed 1000\. Setting this value to 0 causes all new runs to be skipped. The default behavior is to allow only 1 concurrent run.
-                         */
-                        max_concurrent_runs?: number;
-                        /**
-                         * An optional name for the job.
-                         */
-                        name?: string;
-                        /**
-                         * Optional notification settings that are used when sending notifications to each of the `email_notifications` and `webhook_notifications` for this job.
-                         */
-                        notification_settings?: {
-                            /**
-                             * If true, do not send notifications to recipients specified in `on_failure` if the run is canceled.
-                             */
-                            no_alert_for_canceled_runs?: boolean;
-                            /**
-                             * If true, do not send notifications to recipients specified in `on_failure` if the run is skipped.
-                             */
-                            no_alert_for_skipped_runs?: boolean;
-                        };
-                        /**
-                         * Job-level parameter definitions
-                         */
-                        parameters?: {
-                            /**
-                             * Default value of the parameter.
-                             */
-                            default: string;
-                            /**
-                             * The name of the defined parameter. May only contain alphanumeric characters, `_`, `-`, and `.`
-                             */
-                            name: string;
-                        }[];
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                        run_as?: {
-                            /**
-                             * Application ID of an active service principal. Setting this field requires the `servicePrincipal/user` role.
-                             */
-                            service_principal_name?: string;
-                            /**
-                             * The email of an active workspace user. Non-admin users can only set this field to their own email.
-                             */
-                            user_name?: string;
-                        };
-                        /**
-                         * An optional periodic schedule for this job. The default behavior is that the job only runs when triggered by clicking “Run Now” in the Jobs UI or sending an API request to `runNow`.
-                         */
-                        schedule?: {
-                            /**
-                             * Whether this trigger is paused or not.
-                             */
-                            pause_status?: string;
-                            /**
-                             * A Cron expression using Quartz syntax that describes the schedule for a job.
-                             * See [Cron Trigger](http://www.quartz-scheduler.org/documentation/quartz-2.3.0/tutorials/crontrigger.html)
-                             * for details. This field is required."
-                             *
-                             */
-                            quartz_cron_expression: string;
-                            /**
-                             * A Java timezone ID. The schedule for a job is resolved with respect to this timezone.
-                             * See [Java TimeZone](https://docs.oracle.com/javase/7/docs/api/java/util/TimeZone.html) for details.
-                             * This field is required.
-                             *
-                             */
-                            timezone_id: string;
-                        };
-                        /**
-                         * A map of tags associated with the job. These are forwarded to the cluster as cluster tags for jobs clusters, and are subject to the same limitations as cluster tags. A maximum of 25 tags can be added to the job.
-                         */
-                        tags?: {
-                            [k: string]: string;
-                        };
-                        /**
-                         * A list of task specifications to be executed by this job.
-                         */
-                        tasks?: {
-                            /**
-                             * The key of the compute requirement, specified in `job.settings.compute`, to use for execution of this task.
-                             */
-                            compute_key?: string;
-                            /**
-                             * If condition_task, specifies a condition with an outcome that can be used to control the execution of other tasks. Does not require a cluster to execute and does not support retries or notifications.
-                             */
-                            condition_task?: {
-                                /**
-                                 * The left operand of the condition task. Can be either a string value or a job state or parameter reference.
-                                 */
-                                left?: string;
-                                /**
-                                 * * `EQUAL_TO`, `NOT_EQUAL` operators perform string comparison of their operands. This means that `“12.0” == “12”` will evaluate to `false`.
-                                 * * `GREATER_THAN`, `GREATER_THAN_OR_EQUAL`, `LESS_THAN`, `LESS_THAN_OR_EQUAL` operators perform numeric comparison of their operands. `“12.0” >= “12”` will evaluate to `true`, `“10.0” >= “12”` will evaluate to `false`.
-                                 *
-                                 * The boolean comparison to task values can be implemented with operators `EQUAL_TO`, `NOT_EQUAL`. If a task value was set to a boolean value, it will be serialized to `“true”` or `“false”` for the comparison.
-                                 *
-                                 */
-                                op?: string;
-                                /**
-                                 * The right operand of the condition task. Can be either a string value or a job state or parameter reference.
-                                 */
-                                right?: string;
-                            };
-                            /**
-                             * If dbt_task, indicates that this must execute a dbt task. It requires both Databricks SQL and the ability to use a serverless or a pro SQL warehouse.
-                             */
-                            dbt_task?: {
-                                /**
-                                 * Optional name of the catalog to use. The value is the top level in the 3-level namespace of Unity Catalog (catalog / schema / relation). The catalog value can only be specified if a warehouse_id is specified. Requires dbt-databricks >= 1.1.1.
-                                 */
-                                catalog?: string;
-                                /**
-                                 * A list of dbt commands to execute. All commands must start with `dbt`. This parameter must not be empty. A maximum of up to 10 commands can be provided.
-                                 */
-                                commands: string[];
-                                /**
-                                 * Optional (relative) path to the profiles directory. Can only be specified if no warehouse_id is specified. If no warehouse_id is specified and this folder is unset, the root directory is used.
-                                 */
-                                profiles_directory?: string;
-                                /**
-                                 * Optional (relative) path to the project directory, if no value is provided, the root of the git repository is used.
-                                 */
-                                project_directory?: string;
-                                /**
-                                 * Optional schema to write to. This parameter is only used when a warehouse_id is also provided. If not provided, the `default` schema is used.
-                                 */
-                                schema?: string;
-                                /**
-                                 * ID of the SQL warehouse to connect to. If provided, we automatically generate and provide the profile and connection details to dbt. It can be overridden on a per-command basis by using the `--profiles-dir` command line argument.
-                                 */
-                                warehouse_id?: string;
-                            };
-                            /**
-                             * An optional array of objects specifying the dependency graph of the task. All tasks specified in this field must complete successfully before executing this task.
-                             * The key is `task_key`, and the value is the name assigned to the dependent task.
-                             *
-                             */
-                            depends_on?: {
-                                /**
-                                 * Can only be specified on condition task dependencies. The outcome of the dependent task that must be met for this task to run.
-                                 */
-                                outcome?: string;
-                                /**
-                                 * The name of the task this task depends on.
-                                 */
-                                task_key: string;
-                            }[];
-                            /**
-                             * An optional description for this task.
-                             * The maximum length is 4096 bytes.
-                             */
-                            description?: string;
-                            /**
-                             * An optional set of email addresses that is notified when runs of this task begin or complete as well as when this task is deleted. The default behavior is to not send any emails.
-                             */
-                            email_notifications?: {
-                                on_duration_warning_threshold_exceeded?: string[];
-                                /**
-                                 * A list of email addresses to be notified when a run unsuccessfully completes. A run is considered to have completed unsuccessfully if it ends with an `INTERNAL_ERROR` `life_cycle_state` or a `FAILED`, or `TIMED_OUT` result_state. If this is not specified on job creation, reset, or update the list is empty, and notifications are not sent.
-                                 */
-                                on_failure?: string[];
-                                /**
-                                 * A list of email addresses to be notified when a run begins. If not specified on job creation, reset, or update, the list is empty, and notifications are not sent.
-                                 */
-                                on_start?: string[];
-                                /**
-                                 * A list of email addresses to be notified when a run successfully completes. A run is considered to have completed successfully if it ends with a `TERMINATED` `life_cycle_state` and a `SUCCESS` result_state. If not specified on job creation, reset, or update, the list is empty, and notifications are not sent.
-                                 */
-                                on_success?: string[];
-                            };
-                            /**
-                             * If existing_cluster_id, the ID of an existing cluster that is used for all runs of this task. When running tasks on an existing cluster, you may need to manually restart the cluster if it stops responding. We suggest running jobs on new clusters for greater reliability.
-                             */
-                            existing_cluster_id?: string;
-                            health?: {
-                                rules?: {
-                                    metric?: string;
-                                    op?: string;
-                                    value?: number;
-                                }[];
-                            };
-                            /**
-                             * If job_cluster_key, this task is executed reusing the cluster specified in `job.settings.job_clusters`.
-                             */
-                            job_cluster_key?: string;
-                            /**
-                             * An optional list of libraries to be installed on the cluster that executes the task. The default value is an empty list.
-                             */
-                            libraries?: {
-                                /**
-                                 * Specification of a CRAN library to be installed as part of the library
-                                 */
-                                cran?: {
-                                    /**
-                                     * The name of the CRAN package to install.
-                                     */
-                                    package: string;
-                                    /**
-                                     * The repository where the package can be found. If not specified, the default CRAN repo is used.
-                                     */
-                                    repo?: string;
-                                };
-                                /**
-                                 * URI of the egg to be installed. Currently only DBFS and S3 URIs are supported.
-                                 * For example: `{ "egg": "dbfs:/my/egg" }` or
-                                 * `{ "egg": "s3://my-bucket/egg" }`.
-                                 * If S3 is used, please make sure the cluster has read access on the library. You may need to
-                                 * launch the cluster with an IAM role to access the S3 URI.
-                                 */
-                                egg?: string;
-                                /**
-                                 * URI of the jar to be installed. Currently only DBFS and S3 URIs are supported.
-                                 * For example: `{ "jar": "dbfs:/mnt/databricks/library.jar" }` or
-                                 * `{ "jar": "s3://my-bucket/library.jar" }`.
-                                 * If S3 is used, please make sure the cluster has read access on the library. You may need to
-                                 * launch the cluster with an IAM role to access the S3 URI.
-                                 */
-                                jar?: string;
-                                /**
-                                 * Specification of a maven library to be installed. For example:
-                                 * `{ "coordinates": "org.jsoup:jsoup:1.7.2" }`
-                                 */
-                                maven?: {
-                                    /**
-                                     * Gradle-style maven coordinates. For example: "org.jsoup:jsoup:1.7.2".
-                                     */
-                                    coordinates: string;
-                                    /**
-                                     * List of dependences to exclude. For example: `["slf4j:slf4j", "*:hadoop-client"]`.
-                                     *
-                                     * Maven dependency exclusions:
-                                     * https://maven.apache.org/guides/introduction/introduction-to-optional-and-excludes-dependencies.html.
-                                     */
-                                    exclusions?: string[];
-                                    /**
-                                     * Maven repo to install the Maven package from. If omitted, both Maven Central Repository
-                                     * and Spark Packages are searched.
-                                     */
-                                    repo?: string;
-                                };
-                                /**
-                                 * Specification of a PyPi library to be installed. For example:
-                                 * `{ "package": "simplejson" }`
-                                 */
-                                pypi?: {
-                                    /**
-                                     * The name of the pypi package to install. An optional exact version specification is also
-                                     * supported. Examples: "simplejson" and "simplejson==3.8.0".
-                                     */
-                                    package: string;
-                                    /**
-                                     * The repository where the package can be found. If not specified, the default pip index is
-                                     * used.
-                                     */
-                                    repo?: string;
-                                };
-                                /**
-                                 * URI of the wheel to be installed.
-                                 * For example: `{ "whl": "dbfs:/my/whl" }` or `{ "whl": "s3://my-bucket/whl" }`.
-                                 * If S3 is used, please make sure the cluster has read access on the library. You may need to
-                                 * launch the cluster with an IAM role to access the S3 URI.
-                                 */
-                                whl?: string;
-                            }[];
-                            /**
-                             * An optional maximum number of times to retry an unsuccessful run. A run is considered to be unsuccessful if it completes with the `FAILED` result_state or `INTERNAL_ERROR` `life_cycle_state`. The value -1 means to retry indefinitely and the value 0 means to never retry. The default behavior is to never retry.
-                             */
-                            max_retries?: number;
-                            /**
-                             * An optional minimal interval in milliseconds between the start of the failed run and the subsequent retry run. The default behavior is that unsuccessful runs are immediately retried.
-                             */
-                            min_retry_interval_millis?: number;
-                            /**
-                             * If new_cluster, a description of a cluster that is created for only for this task.
-                             */
-                            new_cluster?: {
-                                /**
-                                 * Parameters needed in order to automatically scale clusters up and down based on load.
-                                 * Note: autoscaling works best with DB runtime versions 3.0 or later.
-                                 */
-                                autoscale?: {
-                                    /**
-                                     * The maximum number of workers to which the cluster can scale up when overloaded.
-                                     * Note that `max_workers` must be strictly greater than `min_workers`.
-                                     */
-                                    max_workers: number;
-                                    /**
-                                     * The minimum number of workers to which the cluster can scale down when underutilized.
-                                     * It is also the initial number of workers the cluster will have after creation.
-                                     */
-                                    min_workers: number;
-                                };
-                                /**
-                                 * Automatically terminates the cluster after it is inactive for this time in minutes. If not set,
-                                 * this cluster will not be automatically terminated. If specified, the threshold must be between
-                                 * 10 and 10000 minutes.
-                                 * Users can also set this value to 0 to explicitly disable automatic termination.
-                                 */
-                                autotermination_minutes?: number;
-                                /**
-                                 * Attributes related to clusters running on Amazon Web Services.
-                                 * If not specified at cluster creation, a set of default values will be used.
-                                 */
-                                aws_attributes?: {
-                                    availability?: string;
-                                    /**
-                                     * The number of volumes launched for each instance. Users can choose up to 10 volumes.
-                                     * This feature is only enabled for supported node types. Legacy node types cannot specify
-                                     * custom EBS volumes.
-                                     * For node types with no instance store, at least one EBS volume needs to be specified;
-                                     * otherwise, cluster creation will fail.
-                                     *
-                                     * These EBS volumes will be mounted at `/ebs0`, `/ebs1`, and etc.
-                                     * Instance store volumes will be mounted at `/local_disk0`, `/local_disk1`, and etc.
-                                     *
-                                     * If EBS volumes are attached, Databricks will configure Spark to use only the EBS volumes for
-                                     * scratch storage because heterogenously sized scratch devices can lead to inefficient disk
-                                     * utilization. If no EBS volumes are attached, Databricks will configure Spark to use instance
-                                     * store volumes.
-                                     *
-                                     * Please note that if EBS volumes are specified, then the Spark configuration `spark.local.dir`
-                                     * will be overridden.
-                                     */
-                                    ebs_volume_count?: number;
-                                    /**
-                                     * <needs content added>
-                                     */
-                                    ebs_volume_iops?: number;
-                                    /**
-                                     * The size of each EBS volume (in GiB) launched for each instance. For general purpose
-                                     * SSD, this value must be within the range 100 - 4096. For throughput optimized HDD,
-                                     * this value must be within the range 500 - 4096.
-                                     */
-                                    ebs_volume_size?: number;
-                                    /**
-                                     * <needs content added>
-                                     */
-                                    ebs_volume_throughput?: number;
-                                    ebs_volume_type?: string;
-                                    /**
-                                     * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                                     * If this value is greater than 0, the cluster driver node in particular will be placed on an
-                                     * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                                     * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                                     * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                                     * be placed on `availability` instances. Note that this value does not affect
-                                     * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                                     */
-                                    first_on_demand?: number;
-                                    /**
-                                     * Nodes for this cluster will only be placed on AWS instances with this instance profile. If
-                                     * ommitted, nodes will be placed on instances without an IAM instance profile. The instance
-                                     * profile must have previously been added to the Databricks environment by an account
-                                     * administrator.
-                                     *
-                                     * This feature may only be available to certain customer plans.
-                                     *
-                                     * If this field is ommitted, we will pull in the default from the conf if it exists.
-                                     */
-                                    instance_profile_arn?: string;
-                                    /**
-                                     * The bid price for AWS spot instances, as a percentage of the corresponding instance type's
-                                     * on-demand price.
-                                     * For example, if this field is set to 50, and the cluster needs a new `r3.xlarge` spot
-                                     * instance, then the bid price is half of the price of
-                                     * on-demand `r3.xlarge` instances. Similarly, if this field is set to 200, the bid price is twice
-                                     * the price of on-demand `r3.xlarge` instances. If not specified, the default value is 100.
-                                     * When spot instances are requested for this cluster, only spot instances whose bid price
-                                     * percentage matches this field will be considered.
-                                     * Note that, for safety, we enforce this field to be no more than 10000.
-                                     *
-                                     * The default value and documentation here should be kept consistent with
-                                     * CommonConf.defaultSpotBidPricePercent and CommonConf.maxSpotBidPricePercent.
-                                     */
-                                    spot_bid_price_percent?: number;
-                                    /**
-                                     * Identifier for the availability zone/datacenter in which the cluster resides.
-                                     * This string will be of a form like "us-west-2a". The provided availability
-                                     * zone must be in the same region as the Databricks deployment. For example, "us-west-2a"
-                                     * is not a valid zone id if the Databricks deployment resides in the "us-east-1" region.
-                                     * This is an optional field at cluster creation, and if not specified, a default zone will be used.
-                                     * If the zone specified is "auto", will try to place cluster in a zone with high availability,
-                                     * and will retry placement in a different AZ if there is not enough capacity.
-                                     * See [[AutoAZHelper.scala]] for more details.
-                                     * The list of available zones as well as the default value can be found by using the
-                                     * `List Zones`_ method.
-                                     */
-                                    zone_id?: string;
-                                };
-                                /**
-                                 * Attributes related to clusters running on Microsoft Azure.
-                                 * If not specified at cluster creation, a set of default values will be used.
-                                 */
-                                azure_attributes?: {
-                                    availability?: string;
-                                    /**
-                                     * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                                     * This value should be greater than 0, to make sure the cluster driver node is placed on an
-                                     * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                                     * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                                     * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                                     * be placed on `availability` instances. Note that this value does not affect
-                                     * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                                     */
-                                    first_on_demand?: number;
-                                    /**
-                                     * Defines values necessary to configure and run Azure Log Analytics agent
-                                     */
-                                    log_analytics_info?: {
-                                        /**
-                                         * <needs content added>
-                                         */
-                                        log_analytics_primary_key?: string;
-                                        /**
-                                         * <needs content added>
-                                         */
-                                        log_analytics_workspace_id?: string;
-                                    };
-                                    /**
-                                     * The max bid price to be used for Azure spot instances.
-                                     * The Max price for the bid cannot be higher than the on-demand price of the instance.
-                                     * If not specified, the default value is -1, which specifies that the instance cannot be evicted
-                                     * on the basis of price, and only on the basis of availability. Further, the value should > 0 or -1.
-                                     */
-                                    spot_bid_max_price?: number;
-                                };
-                                /**
-                                 * The configuration for delivering spark logs to a long-term storage destination.
-                                 * Two kinds of destinations (dbfs and s3) are supported. Only one destination can be specified
-                                 * for one cluster. If the conf is given, the logs will be delivered to the destination every
-                                 * `5 mins`. The destination of driver logs is `$destination/$clusterId/driver`, while
-                                 * the destination of executor logs is `$destination/$clusterId/executor`.
-                                 */
-                                cluster_log_conf?: {
-                                    /**
-                                     * destination needs to be provided. e.g.
-                                     * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                                     */
-                                    dbfs?: {
-                                        /**
-                                         * dbfs destination, e.g. `dbfs:/my/path`
-                                         */
-                                        destination?: string;
-                                    };
-                                    /**
-                                     * destination and either the region or endpoint need to be provided. e.g.
-                                     * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                                     * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                                     * `instance_profile_arn` has permission to write data to the s3 destination.
-                                     */
-                                    s3?: {
-                                        /**
-                                         * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                                         * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                                         * the destination bucket and prefix. The full list of possible canned acl can be found at
-                                         * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                                         * Please also note that by default only the object owner gets full controls. If you are using cross account
-                                         * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                                         * read the logs.
-                                         */
-                                        canned_acl?: string;
-                                        /**
-                                         * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                                         * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                                         * destination. Please also note that you cannot use AWS keys to deliver logs.
-                                         */
-                                        destination?: string;
-                                        /**
-                                         * (Optional) Flag to enable server side encryption, `false` by default.
-                                         */
-                                        enable_encryption?: boolean;
-                                        /**
-                                         * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                                         * encryption is enabled and the default type is `sse-s3`.
-                                         */
-                                        encryption_type?: string;
-                                        /**
-                                         * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                                         * If both are set, endpoint will be used.
-                                         */
-                                        endpoint?: string;
-                                        /**
-                                         * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                                         */
-                                        kms_key?: string;
-                                        /**
-                                         * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                                         * endpoint will be used.
-                                         */
-                                        region?: string;
-                                    };
-                                };
-                                /**
-                                 * Cluster name requested by the user. This doesn't have to be unique.
-                                 * If not specified at creation, the cluster name will be an empty string.
-                                 *
-                                 */
-                                cluster_name?: string;
-                                cluster_source?: string;
-                                /**
-                                 * Additional tags for cluster resources. Databricks will tag all cluster resources (e.g., AWS
-                                 * instances and EBS volumes) with these tags in addition to `default_tags`. Notes:
-                                 *
-                                 * - Currently, Databricks allows at most 45 custom tags
-                                 *
-                                 * - Clusters can only reuse cloud resources if the resources' tags are a subset of the cluster tags
-                                 */
-                                custom_tags?: {
-                                    [k: string]: string;
-                                };
-                                data_security_mode?: string;
-                                docker_image?: {
-                                    basic_auth?: {
-                                        /**
-                                         * Password of the user
-                                         */
-                                        password?: string;
-                                        /**
-                                         * Name of the user
-                                         */
-                                        username?: string;
-                                    };
-                                    /**
-                                     * URL of the docker image.
-                                     */
-                                    url?: string;
-                                };
-                                /**
-                                 * The optional ID of the instance pool for the driver of the cluster belongs.
-                                 * The pool cluster uses the instance pool with id (instance_pool_id) if the driver pool is not
-                                 * assigned.
-                                 */
-                                driver_instance_pool_id?: string;
-                                /**
-                                 * The node type of the Spark driver. Note that this field is optional;
-                                 * if unset, the driver node type will be set as the same value
-                                 * as `node_type_id` defined above.
-                                 *
-                                 */
-                                driver_node_type_id?: string;
-                                /**
-                                 * Autoscaling Local Storage: when enabled, this cluster will dynamically acquire additional disk
-                                 * space when its Spark workers are running low on disk space. This feature requires specific AWS
-                                 * permissions to function correctly - refer to the User Guide for more details.
-                                 */
-                                enable_elastic_disk?: boolean;
-                                /**
-                                 * Whether to enable LUKS on cluster VMs' local disks
-                                 */
-                                enable_local_disk_encryption?: boolean;
-                                /**
-                                 * Attributes related to clusters running on Google Cloud Platform.
-                                 * If not specified at cluster creation, a set of default values will be used.
-                                 */
-                                gcp_attributes?: {
-                                    availability?: string;
-                                    /**
-                                     * boot disk size in GB
-                                     */
-                                    boot_disk_size?: number;
-                                    /**
-                                     * If provided, the cluster will impersonate the google service account when accessing
-                                     * gcloud services (like GCS). The google service account
-                                     * must have previously been added to the Databricks environment by an account
-                                     * administrator.
-                                     */
-                                    google_service_account?: string;
-                                    /**
-                                     * If provided, each node (workers and driver) in the cluster will have this number of local SSDs attached. Each local SSD is 375GB in size. Refer to [GCP documentation](https://cloud.google.com/compute/docs/disks/local-ssd#choose_number_local_ssds) for the supported number of local SSDs for each instance type.
-                                     */
-                                    local_ssd_count?: number;
-                                };
-                                /**
-                                 * The configuration for storing init scripts. Any number of destinations can be specified. The scripts are executed sequentially in the order provided. If `cluster_log_conf` is specified, init script logs are sent to `<destination>/<cluster-ID>/init_scripts`.
-                                 */
-                                init_scripts?: {
-                                    /**
-                                     * destination needs to be provided. e.g.
-                                     * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                                     */
-                                    dbfs?: {
-                                        /**
-                                         * dbfs destination, e.g. `dbfs:/my/path`
-                                         */
-                                        destination?: string;
-                                    };
-                                    /**
-                                     * destination and either the region or endpoint need to be provided. e.g.
-                                     * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                                     * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                                     * `instance_profile_arn` has permission to write data to the s3 destination.
-                                     */
-                                    s3?: {
-                                        /**
-                                         * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                                         * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                                         * the destination bucket and prefix. The full list of possible canned acl can be found at
-                                         * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                                         * Please also note that by default only the object owner gets full controls. If you are using cross account
-                                         * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                                         * read the logs.
-                                         */
-                                        canned_acl?: string;
-                                        /**
-                                         * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                                         * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                                         * destination. Please also note that you cannot use AWS keys to deliver logs.
-                                         */
-                                        destination?: string;
-                                        /**
-                                         * (Optional) Flag to enable server side encryption, `false` by default.
-                                         */
-                                        enable_encryption?: boolean;
-                                        /**
-                                         * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                                         * encryption is enabled and the default type is `sse-s3`.
-                                         */
-                                        encryption_type?: string;
-                                        /**
-                                         * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                                         * If both are set, endpoint will be used.
-                                         */
-                                        endpoint?: string;
-                                        /**
-                                         * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                                         */
-                                        kms_key?: string;
-                                        /**
-                                         * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                                         * endpoint will be used.
-                                         */
-                                        region?: string;
-                                    };
-                                    volumes?: {
-                                        destination?: string;
-                                    };
-                                    /**
-                                     * destination needs to be provided. e.g.
-                                     * `{ "workspace" : { "destination" : "/Users/user1@databricks.com/my-init.sh" } }`
-                                     */
-                                    workspace?: {
-                                        /**
-                                         * workspace files destination, e.g. `/Users/user1@databricks.com/my-init.sh`
-                                         */
-                                        destination?: string;
-                                    };
-                                }[];
-                                /**
-                                 * The optional ID of the instance pool to which the cluster belongs.
-                                 */
-                                instance_pool_id?: string;
-                                /**
-                                 * This field encodes, through a single value, the resources available to each of
-                                 * the Spark nodes in this cluster. For example, the Spark nodes can be provisioned
-                                 * and optimized for memory or compute intensive workloads. A list of available node
-                                 * types can be retrieved by using the :method:clusters/listNodeTypes API call.
-                                 *
-                                 */
-                                node_type_id?: string;
-                                /**
-                                 * Number of worker nodes that this cluster should have. A cluster has one Spark Driver
-                                 * and `num_workers` Executors for a total of `num_workers` + 1 Spark nodes.
-                                 *
-                                 * Note: When reading the properties of a cluster, this field reflects the desired number
-                                 * of workers rather than the actual current number of workers. For instance, if a cluster
-                                 * is resized from 5 to 10 workers, this field will immediately be updated to reflect
-                                 * the target size of 10 workers, whereas the workers listed in `spark_info` will gradually
-                                 * increase from 5 to 10 as the new nodes are provisioned.
-                                 */
-                                num_workers?: number;
-                                /**
-                                 * The ID of the cluster policy used to create the cluster if applicable.
-                                 */
-                                policy_id?: string;
-                                runtime_engine?: string;
-                                /**
-                                 * Single user name if data_security_mode is `SINGLE_USER`
-                                 */
-                                single_user_name?: string;
-                                /**
-                                 * An object containing a set of optional, user-specified Spark configuration key-value pairs.
-                                 * Users can also pass in a string of extra JVM options to the driver and the executors via
-                                 * `spark.driver.extraJavaOptions` and `spark.executor.extraJavaOptions` respectively.
-                                 *
-                                 */
-                                spark_conf?: {
-                                    [k: string]: string;
-                                };
-                                /**
-                                 * An object containing a set of optional, user-specified environment variable key-value pairs.
-                                 * Please note that key-value pair of the form (X,Y) will be exported as is (i.e.,
-                                 * `export X='Y'`) while launching the driver and workers.
-                                 *
-                                 * In order to specify an additional set of `SPARK_DAEMON_JAVA_OPTS`, we recommend appending
-                                 * them to `$SPARK_DAEMON_JAVA_OPTS` as shown in the example below. This ensures that all
-                                 * default databricks managed environmental variables are included as well.
-                                 *
-                                 * Example Spark environment variables:
-                                 * `{"SPARK_WORKER_MEMORY": "28000m", "SPARK_LOCAL_DIRS": "/local_disk0"}` or
-                                 * `{"SPARK_DAEMON_JAVA_OPTS": "$SPARK_DAEMON_JAVA_OPTS -Dspark.shuffle.service.enabled=true"}`
-                                 */
-                                spark_env_vars?: {
-                                    [k: string]: string;
-                                };
-                                /**
-                                 * The Spark version of the cluster, e.g. `3.3.x-scala2.11`.
-                                 * A list of available Spark versions can be retrieved by using
-                                 * the :method:clusters/sparkVersions API call.
-                                 *
-                                 */
-                                spark_version?: string;
-                                /**
-                                 * SSH public key contents that will be added to each Spark node in this cluster. The
-                                 * corresponding private keys can be used to login with the user name `ubuntu` on port `2200`.
-                                 * Up to 10 keys can be specified.
-                                 */
-                                ssh_public_keys?: string[];
-                                workload_type?: {
-                                    /**
-                                     *  defined what type of clients can use the cluster. E.g. Notebooks, Jobs
-                                     */
-                                    clients?: {
-                                        /**
-                                         * With jobs set, the cluster can be used for jobs
-                                         */
-                                        jobs?: boolean;
-                                        /**
-                                         * With notebooks set, this cluster can be used for notebooks
-                                         */
-                                        notebooks?: boolean;
-                                    };
-                                };
-                            };
-                            /**
-                             * If notebook_task, indicates that this task must run a notebook. This field may not be specified in conjunction with spark_jar_task.
-                             */
-                            notebook_task?: {
-                                /**
-                                 * Base parameters to be used for each run of this job. If the run is initiated by a call to
-                                 * :method:jobs/runNow with parameters specified, the two parameters maps are merged. If the same key is specified in
-                                 * `base_parameters` and in `run-now`, the value from `run-now` is used.
-                                 *
-                                 * Use [Task parameter variables](https://docs.databricks.com/jobs.html#parameter-variables) to set parameters containing information about job runs.
-                                 *
-                                 * If the notebook takes a parameter that is not specified in the job’s `base_parameters` or the `run-now` override parameters,
-                                 * the default value from the notebook is used.
-                                 *
-                                 * Retrieve these parameters in a notebook using [dbutils.widgets.get](https://docs.databricks.com/dev-tools/databricks-utils.html#dbutils-widgets).
-                                 *
-                                 */
-                                base_parameters?: {
-                                    [k: string]: string;
-                                };
-                                /**
-                                 * The path of the notebook to be run in the Databricks workspace or remote repository.
-                                 * For notebooks stored in the Databricks workspace, the path must be absolute and begin with a slash.
-                                 * For notebooks stored in a remote repository, the path must be relative. This field is required.
-                                 *
-                                 */
-                                notebook_path: string;
-                                /**
-                                 * Optional location type of the Python file. When set to `WORKSPACE` or not specified, the file will be retrieved
-                                 * from the local <Databricks> workspace or cloud location (if the `python_file` has a URI format). When set to `GIT`,
-                                 * the Python file will be retrieved from a Git repository defined in `git_source`.
-                                 *
-                                 * * `WORKSPACE`: The Python file is located in a <Databricks> workspace or at a cloud filesystem URI.
-                                 * * `GIT`: The Python file is located in a remote Git repository.
-                                 *
-                                 */
-                                source?: string;
-                            };
-                            /**
-                             * Optional notification settings that are used when sending notifications to each of the `email_notifications` for this task.
-                             */
-                            notification_settings?: {
-                                /**
-                                 * If true, do not send notifications to recipients specified in `on_start` for the retried runs and do not send notifications to recipients specified in `on_failure` until the last retry of the run.
-                                 */
-                                alert_on_last_attempt?: boolean;
-                                /**
-                                 * If true, do not send notifications to recipients specified in `on_failure` if the run is canceled.
-                                 */
-                                no_alert_for_canceled_runs?: boolean;
-                                /**
-                                 * If true, do not send notifications to recipients specified in `on_failure` if the run is skipped.
-                                 */
-                                no_alert_for_skipped_runs?: boolean;
-                            };
-                            /**
-                             * If pipeline_task, indicates that this task must execute a Pipeline.
-                             */
-                            pipeline_task?: {
-                                /**
-                                 * If true, a full refresh will be triggered on the delta live table.
-                                 */
-                                full_refresh?: boolean;
-                                /**
-                                 * The full name of the pipeline task to execute.
-                                 */
-                                pipeline_id?: string;
-                            };
-                            /**
-                             * If python_wheel_task, indicates that this job must execute a PythonWheel.
-                             */
-                            python_wheel_task?: {
-                                /**
-                                 * Named entry point to use, if it does not exist in the metadata of the package it executes the function from the package directly using `$packageName.$entryPoint()`
-                                 */
-                                entry_point?: string;
-                                /**
-                                 * Command-line parameters passed to Python wheel task in the form of `["--name=task", "--data=dbfs:/path/to/data.json"]`. Leave it empty if `parameters` is not null.
-                                 */
-                                named_parameters?: {
-                                    [k: string]: string;
-                                };
-                                /**
-                                 * Name of the package to execute
-                                 */
-                                package_name?: string;
-                                /**
-                                 * Command-line parameters passed to Python wheel task. Leave it empty if `named_parameters` is not null.
-                                 */
-                                parameters?: string[];
-                            };
-                            /**
-                             * An optional policy to specify whether to retry a task when it times out. The default behavior is to not retry on timeout.
-                             */
-                            retry_on_timeout?: boolean;
-                            /**
-                             * An optional value specifying the condition determining whether the task is run once its dependencies have been completed. When omitted, defaults to `ALL_SUCCESS`.
-                             *
-                             * * `ALL_SUCCESS`: All dependencies have executed and succeeded
-                             * * `AT_LEAST_ONE_SUCCESS`: At least one dependency has succeeded
-                             * * `NONE_FAILED`: None of the dependencies have failed and at least one was executed
-                             * * `ALL_DONE`: All dependencies completed and at least one was executed
-                             * * `AT_LEAST_ONE_FAILED`: At least one dependency failed
-                             * * `ALL_FAILED`: ALl dependencies have failed
-                             *
-                             */
-                            run_if?: string;
-                            run_job_task?: {
-                                job_id: number;
-                                job_parameters?: unknown;
-                            };
-                            /**
-                             * If spark_jar_task, indicates that this task must run a JAR.
-                             */
-                            spark_jar_task?: {
-                                /**
-                                 * Deprecated since 04/2016. Provide a `jar` through the `libraries` field instead. For an example, see :method:jobs/create.
-                                 *
-                                 */
-                                jar_uri?: string;
-                                /**
-                                 * The full name of the class containing the main method to be executed. This class must be contained in a JAR provided as a library.
-                                 *
-                                 * The code must use `SparkContext.getOrCreate` to obtain a Spark context; otherwise, runs of the job fail.
-                                 */
-                                main_class_name?: string;
-                                /**
-                                 * Parameters passed to the main method.
-                                 *
-                                 * Use [Task parameter variables](https://docs.databricks.com/jobs.html#parameter-variables) to set parameters containing information about job runs.
-                                 *
-                                 */
-                                parameters?: string[];
-                            };
-                            /**
-                             * If spark_python_task, indicates that this task must run a Python file.
-                             */
-                            spark_python_task?: {
-                                /**
-                                 * Command line parameters passed to the Python file.
-                                 *
-                                 * Use [Task parameter variables](https://docs.databricks.com/jobs.html#parameter-variables) to set parameters containing information about job runs.
-                                 *
-                                 */
-                                parameters?: string[];
-                                /**
-                                 * The Python file to be executed. Cloud file URIs (such as dbfs:/, s3:/, adls:/, gcs:/) and workspace paths are supported. For python files stored in the Databricks workspace, the path must be absolute and begin with `/`. For files stored in a remote repository, the path must be relative. This field is required.
-                                 */
-                                python_file: string;
-                                /**
-                                 * Optional location type of the Python file. When set to `WORKSPACE` or not specified, the file will be retrieved
-                                 * from the local <Databricks> workspace or cloud location (if the `python_file` has a URI format). When set to `GIT`,
-                                 * the Python file will be retrieved from a Git repository defined in `git_source`.
-                                 *
-                                 * * `WORKSPACE`: The Python file is located in a <Databricks> workspace or at a cloud filesystem URI.
-                                 * * `GIT`: The Python file is located in a remote Git repository.
-                                 *
-                                 */
-                                source?: string;
-                            };
-                            /**
-                             * If spark_submit_task, indicates that this task must be launched by the spark submit script. This task can run only on new clusters.
-                             */
-                            spark_submit_task?: {
-                                /**
-                                 * Command-line parameters passed to spark submit.
-                                 *
-                                 * Use [Task parameter variables](https://docs.databricks.com/jobs.html#parameter-variables) to set parameters containing information about job runs.
-                                 *
-                                 */
-                                parameters?: string[];
-                            };
-                            /**
-                             * If sql_task, indicates that this job must execute a SQL task.
-                             */
-                            sql_task?: {
-                                /**
-                                 * If alert, indicates that this job must refresh a SQL alert.
-                                 */
-                                alert?: {
-                                    /**
-                                     * The canonical identifier of the SQL alert.
-                                     */
-                                    alert_id: string;
-                                    /**
-                                     * If true, the alert notifications are not sent to subscribers.
-                                     */
-                                    pause_subscriptions?: boolean;
-                                    /**
-                                     * If specified, alert notifications are sent to subscribers.
-                                     */
-                                    subscriptions?: {
-                                        /**
-                                         * The canonical identifier of the destination to receive email notification.
-                                         */
-                                        destination_id?: string;
-                                        /**
-                                         * The user name to receive the subscription email.
-                                         */
-                                        user_name?: string;
-                                    }[];
-                                };
-                                /**
-                                 * If dashboard, indicates that this job must refresh a SQL dashboard.
-                                 */
-                                dashboard?: {
-                                    /**
-                                     * Subject of the email sent to subscribers of this task.
-                                     */
-                                    custom_subject?: string;
-                                    /**
-                                     * The canonical identifier of the SQL dashboard.
-                                     */
-                                    dashboard_id: string;
-                                    /**
-                                     * If true, the dashboard snapshot is not taken, and emails are not sent to subscribers.
-                                     */
-                                    pause_subscriptions?: boolean;
-                                    /**
-                                     * If specified, dashboard snapshots are sent to subscriptions.
-                                     */
-                                    subscriptions?: {
-                                        /**
-                                         * The canonical identifier of the destination to receive email notification.
-                                         */
-                                        destination_id?: string;
-                                        /**
-                                         * The user name to receive the subscription email.
-                                         */
-                                        user_name?: string;
-                                    }[];
-                                };
-                                /**
-                                 * If file, indicates that this job runs a SQL file in a remote Git repository. Only one SQL statement is supported in a file. Multiple SQL statements separated by semicolons (;) are not permitted.
-                                 */
-                                file?: {
-                                    /**
-                                     * Relative path of the SQL file in the remote Git repository.
-                                     */
-                                    path: string;
-                                };
-                                /**
-                                 * Parameters to be used for each run of this job. The SQL alert task does not support custom parameters.
-                                 */
-                                parameters?: {
-                                    [k: string]: string;
-                                };
-                                /**
-                                 * If query, indicates that this job must execute a SQL query.
-                                 */
-                                query?: {
-                                    /**
-                                     * The canonical identifier of the SQL query.
-                                     */
-                                    query_id: string;
-                                };
-                                /**
-                                 * The canonical identifier of the SQL warehouse. Only serverless and pro SQL warehouses are supported.
-                                 */
-                                warehouse_id: string;
-                            };
-                            /**
-                             * A unique name for the task. This field is used to refer to this task from other tasks.
-                             * This field is required and must be unique within its parent job.
-                             * On Update or Reset, this field is used to reference the tasks to be updated or reset.
-                             * The maximum length is 100 characters.
-                             */
-                            task_key: string;
-                            /**
-                             * An optional timeout applied to each run of this job task. The default behavior is to have no timeout.
-                             */
-                            timeout_seconds?: number;
-                        }[];
-                        /**
-                         * An optional timeout applied to each run of this job. The default behavior is to have no timeout.
-                         */
-                        timeout_seconds?: number;
-                        /**
-                         * Trigger settings for the job. Can be used to trigger a run when new files arrive in an external location. The default behavior is that the job runs only when triggered by clicking “Run Now” in the Jobs UI or sending an API request to `runNow`.
-                         */
-                        trigger?: {
-                            /**
-                             * File arrival trigger settings.
-                             */
-                            file_arrival?: {
-                                /**
-                                 * If set, the trigger starts a run only after the specified amount of time passed since
-                                 * the last time the trigger fired. The minimum allowed value is 60 seconds
-                                 *
-                                 */
-                                min_time_between_triggers_seconds?: number;
-                                /**
-                                 * URL to be monitored for file arrivals. The path must point to the root or a subpath of the external location.
-                                 */
-                                url?: string;
-                                /**
-                                 * If set, the trigger starts a run only after no file activity has occurred for the specified amount of time.
-                                 * This makes it possible to wait for a batch of incoming files to arrive before triggering a run. The
-                                 * minimum allowed value is 60 seconds.
-                                 *
-                                 */
-                                wait_after_last_change_seconds?: number;
-                            };
-                            /**
-                             * Whether this trigger is paused or not.
-                             */
-                            pause_status?: string;
-                        };
-                        /**
-                         * A collection of system notification IDs to notify when the run begins or completes. The default behavior is to not send any system notifications.
-                         */
-                        webhook_notifications?: {
-                            on_duration_warning_threshold_exceeded?: {
-                                id?: string;
-                            }[];
-                            /**
-                             * An optional list of system notification IDs to call when the run fails. A maximum of 3 destinations can be specified for the `on_failure` property.
-                             */
-                            on_failure?: {
-                                id?: string;
-                            }[];
-                            /**
-                             * An optional list of system notification IDs to call when the run starts. A maximum of 3 destinations can be specified for the `on_start` property.
-                             */
-                            on_start?: {
-                                id?: string;
-                            }[];
-                            /**
-                             * An optional list of system notification IDs to call when the run completes successfully. A maximum of 3 destinations can be specified for the `on_success` property.
-                             */
-                            on_success?: {
-                                id?: string;
-                            }[];
-                        };
-                    };
-                };
-                model_serving_endpoints?: {
-                    [k: string]: {
-                        config: {
-                            served_models: {
-                                environment_vars?: {
-                                    [k: string]: string;
-                                };
-                                instance_profile_arn?: string;
-                                model_name: string;
-                                model_version: string;
-                                name?: string;
-                                scale_to_zero_enabled: boolean;
-                                workload_size: string;
-                            }[];
-                            traffic_config?: {
-                                routes?: {
-                                    served_model_name: string;
-                                    traffic_percentage: number;
-                                }[];
-                            };
-                        };
-                        name: string;
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                    };
-                };
-                /**
-                 * List of MLflow models
-                 */
-                models?: {
-                    [k: string]: {
-                        /**
-                         * Timestamp recorded when this `registered_model` was created.
-                         */
-                        creation_timestamp?: number;
-                        /**
-                         * Description of this `registered_model`.
-                         */
-                        description?: string;
-                        /**
-                         * Timestamp recorded when metadata for this `registered_model` was last updated.
-                         */
-                        last_updated_timestamp?: number;
-                        /**
-                         * Collection of latest model versions for each stage.
-                         * Only contains models with current `READY` status.
-                         */
-                        latest_versions?: {
-                            /**
-                             * Timestamp recorded when this `model_version` was created.
-                             */
-                            creation_timestamp?: number;
-                            /**
-                             * Current stage for this `model_version`.
-                             */
-                            current_stage?: string;
-                            /**
-                             * Description of this `model_version`.
-                             */
-                            description?: string;
-                            /**
-                             * Timestamp recorded when metadata for this `model_version` was last updated.
-                             */
-                            last_updated_timestamp?: number;
-                            /**
-                             * Unique name of the model
-                             */
-                            name?: string;
-                            /**
-                             * MLflow run ID used when creating `model_version`, if `source` was generated by an
-                             * experiment run stored in MLflow tracking server.
-                             */
-                            run_id?: string;
-                            /**
-                             * Run Link: Direct link to the run that generated this version
-                             */
-                            run_link?: string;
-                            /**
-                             * URI indicating the location of the source model artifacts, used when creating `model_version`
-                             */
-                            source?: string;
-                            /**
-                             * Current status of `model_version`
-                             */
-                            status?: string;
-                            /**
-                             * Details on current `status`, if it is pending or failed.
-                             */
-                            status_message?: string;
-                            /**
-                             * Tags: Additional metadata key-value pairs for this `model_version`.
-                             */
-                            tags?: {
-                                /**
-                                 * The tag key.
-                                 */
-                                key?: string;
-                                /**
-                                 * The tag value.
-                                 */
-                                value?: string;
-                            }[];
-                            /**
-                             * User that created this `model_version`.
-                             */
-                            user_id?: string;
-                            /**
-                             * Model's version number.
-                             */
-                            version?: string;
-                        }[];
-                        /**
-                         * Unique name for the model.
-                         */
-                        name?: string;
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                        /**
-                         * Tags: Additional metadata key-value pairs for this `registered_model`.
-                         */
-                        tags?: {
-                            /**
-                             * The tag key.
-                             */
-                            key?: string;
-                            /**
-                             * The tag value.
-                             */
-                            value?: string;
-                        }[];
-                        /**
-                         * User that created this `registered_model`
-                         */
-                        user_id?: string;
-                    };
-                };
-                /**
-                 * List of DLT pipelines
-                 */
-                pipelines?: {
-                    [k: string]: {
-                        /**
-                         * A catalog in Unity Catalog to publish data from this pipeline to. If `target` is specified, tables in this pipeline are published to a `target` schema inside `catalog` (for example, `catalog`.`target`.`table`). If `target` is not specified, no data is published to Unity Catalog.
-                         */
-                        catalog?: string;
-                        /**
-                         * DLT Release Channel that specifies which version to use.
-                         */
-                        channel?: string;
-                        /**
-                         * Cluster settings for this pipeline deployment.
-                         */
-                        clusters?: {
-                            /**
-                             * Note: This field won't be persisted. Only API users will check this field.
-                             */
-                            apply_policy_default_values?: boolean;
-                            /**
-                             * Parameters needed in order to automatically scale clusters up and down based on load.
-                             * Note: autoscaling works best with DB runtime versions 3.0 or later.
-                             */
-                            autoscale?: {
-                                /**
-                                 * The maximum number of workers to which the cluster can scale up when overloaded.
-                                 * Note that `max_workers` must be strictly greater than `min_workers`.
-                                 */
-                                max_workers: number;
-                                /**
-                                 * The minimum number of workers to which the cluster can scale down when underutilized.
-                                 * It is also the initial number of workers the cluster will have after creation.
-                                 */
-                                min_workers: number;
-                            };
-                            /**
-                             * Attributes related to clusters running on Amazon Web Services.
-                             * If not specified at cluster creation, a set of default values will be used.
-                             */
-                            aws_attributes?: {
-                                availability?: string;
-                                /**
-                                 * The number of volumes launched for each instance. Users can choose up to 10 volumes.
-                                 * This feature is only enabled for supported node types. Legacy node types cannot specify
-                                 * custom EBS volumes.
-                                 * For node types with no instance store, at least one EBS volume needs to be specified;
-                                 * otherwise, cluster creation will fail.
-                                 *
-                                 * These EBS volumes will be mounted at `/ebs0`, `/ebs1`, and etc.
-                                 * Instance store volumes will be mounted at `/local_disk0`, `/local_disk1`, and etc.
-                                 *
-                                 * If EBS volumes are attached, Databricks will configure Spark to use only the EBS volumes for
-                                 * scratch storage because heterogenously sized scratch devices can lead to inefficient disk
-                                 * utilization. If no EBS volumes are attached, Databricks will configure Spark to use instance
-                                 * store volumes.
-                                 *
-                                 * Please note that if EBS volumes are specified, then the Spark configuration `spark.local.dir`
-                                 * will be overridden.
-                                 */
-                                ebs_volume_count?: number;
-                                /**
-                                 * <needs content added>
-                                 */
-                                ebs_volume_iops?: number;
-                                /**
-                                 * The size of each EBS volume (in GiB) launched for each instance. For general purpose
-                                 * SSD, this value must be within the range 100 - 4096. For throughput optimized HDD,
-                                 * this value must be within the range 500 - 4096.
-                                 */
-                                ebs_volume_size?: number;
-                                /**
-                                 * <needs content added>
-                                 */
-                                ebs_volume_throughput?: number;
-                                ebs_volume_type?: string;
-                                /**
-                                 * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                                 * If this value is greater than 0, the cluster driver node in particular will be placed on an
-                                 * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                                 * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                                 * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                                 * be placed on `availability` instances. Note that this value does not affect
-                                 * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                                 */
-                                first_on_demand?: number;
-                                /**
-                                 * Nodes for this cluster will only be placed on AWS instances with this instance profile. If
-                                 * ommitted, nodes will be placed on instances without an IAM instance profile. The instance
-                                 * profile must have previously been added to the Databricks environment by an account
-                                 * administrator.
-                                 *
-                                 * This feature may only be available to certain customer plans.
-                                 *
-                                 * If this field is ommitted, we will pull in the default from the conf if it exists.
-                                 */
-                                instance_profile_arn?: string;
-                                /**
-                                 * The bid price for AWS spot instances, as a percentage of the corresponding instance type's
-                                 * on-demand price.
-                                 * For example, if this field is set to 50, and the cluster needs a new `r3.xlarge` spot
-                                 * instance, then the bid price is half of the price of
-                                 * on-demand `r3.xlarge` instances. Similarly, if this field is set to 200, the bid price is twice
-                                 * the price of on-demand `r3.xlarge` instances. If not specified, the default value is 100.
-                                 * When spot instances are requested for this cluster, only spot instances whose bid price
-                                 * percentage matches this field will be considered.
-                                 * Note that, for safety, we enforce this field to be no more than 10000.
-                                 *
-                                 * The default value and documentation here should be kept consistent with
-                                 * CommonConf.defaultSpotBidPricePercent and CommonConf.maxSpotBidPricePercent.
-                                 */
-                                spot_bid_price_percent?: number;
-                                /**
-                                 * Identifier for the availability zone/datacenter in which the cluster resides.
-                                 * This string will be of a form like "us-west-2a". The provided availability
-                                 * zone must be in the same region as the Databricks deployment. For example, "us-west-2a"
-                                 * is not a valid zone id if the Databricks deployment resides in the "us-east-1" region.
-                                 * This is an optional field at cluster creation, and if not specified, a default zone will be used.
-                                 * If the zone specified is "auto", will try to place cluster in a zone with high availability,
-                                 * and will retry placement in a different AZ if there is not enough capacity.
-                                 * See [[AutoAZHelper.scala]] for more details.
-                                 * The list of available zones as well as the default value can be found by using the
-                                 * `List Zones`_ method.
-                                 */
-                                zone_id?: string;
-                            };
-                            /**
-                             * Attributes related to clusters running on Microsoft Azure.
-                             * If not specified at cluster creation, a set of default values will be used.
-                             */
-                            azure_attributes?: {
-                                availability?: string;
-                                /**
-                                 * The first `first_on_demand` nodes of the cluster will be placed on on-demand instances.
-                                 * This value should be greater than 0, to make sure the cluster driver node is placed on an
-                                 * on-demand instance. If this value is greater than or equal to the current cluster size, all
-                                 * nodes will be placed on on-demand instances. If this value is less than the current cluster
-                                 * size, `first_on_demand` nodes will be placed on on-demand instances and the remainder will
-                                 * be placed on `availability` instances. Note that this value does not affect
-                                 * cluster size and cannot currently be mutated over the lifetime of a cluster.
-                                 */
-                                first_on_demand?: number;
-                                /**
-                                 * Defines values necessary to configure and run Azure Log Analytics agent
-                                 */
-                                log_analytics_info?: {
-                                    /**
-                                     * <needs content added>
-                                     */
-                                    log_analytics_primary_key?: string;
-                                    /**
-                                     * <needs content added>
-                                     */
-                                    log_analytics_workspace_id?: string;
-                                };
-                                /**
-                                 * The max bid price to be used for Azure spot instances.
-                                 * The Max price for the bid cannot be higher than the on-demand price of the instance.
-                                 * If not specified, the default value is -1, which specifies that the instance cannot be evicted
-                                 * on the basis of price, and only on the basis of availability. Further, the value should > 0 or -1.
-                                 */
-                                spot_bid_max_price?: number;
-                            };
-                            /**
-                             * The configuration for delivering spark logs to a long-term storage destination.
-                             * Only dbfs destinations are supported. Only one destination can be specified
-                             * for one cluster. If the conf is given, the logs will be delivered to the destination every
-                             * `5 mins`. The destination of driver logs is `$destination/$clusterId/driver`, while
-                             * the destination of executor logs is `$destination/$clusterId/executor`.
-                             *
-                             */
-                            cluster_log_conf?: {
-                                /**
-                                 * destination needs to be provided. e.g.
-                                 * `{ "dbfs" : { "destination" : "dbfs:/home/cluster_log" } }`
-                                 */
-                                dbfs?: {
-                                    /**
-                                     * dbfs destination, e.g. `dbfs:/my/path`
-                                     */
-                                    destination?: string;
-                                };
-                                /**
-                                 * destination and either the region or endpoint need to be provided. e.g.
-                                 * `{ "s3": { "destination" : "s3://cluster_log_bucket/prefix", "region" : "us-west-2" } }`
-                                 * Cluster iam role is used to access s3, please make sure the cluster iam role in
-                                 * `instance_profile_arn` has permission to write data to the s3 destination.
-                                 */
-                                s3?: {
-                                    /**
-                                     * (Optional) Set canned access control list for the logs, e.g. `bucket-owner-full-control`.
-                                     * If `canned_cal` is set, please make sure the cluster iam role has `s3:PutObjectAcl` permission on
-                                     * the destination bucket and prefix. The full list of possible canned acl can be found at
-                                     * http://docs.aws.amazon.com/AmazonS3/latest/dev/acl-overview.html#canned-acl.
-                                     * Please also note that by default only the object owner gets full controls. If you are using cross account
-                                     * role for writing data, you may want to set `bucket-owner-full-control` to make bucket owner able to
-                                     * read the logs.
-                                     */
-                                    canned_acl?: string;
-                                    /**
-                                     * S3 destination, e.g. `s3://my-bucket/some-prefix` Note that logs will be delivered using
-                                     * cluster iam role, please make sure you set cluster iam role and the role has write access to the
-                                     * destination. Please also note that you cannot use AWS keys to deliver logs.
-                                     */
-                                    destination?: string;
-                                    /**
-                                     * (Optional) Flag to enable server side encryption, `false` by default.
-                                     */
-                                    enable_encryption?: boolean;
-                                    /**
-                                     * (Optional) The encryption type, it could be `sse-s3` or `sse-kms`. It will be used only when
-                                     * encryption is enabled and the default type is `sse-s3`.
-                                     */
-                                    encryption_type?: string;
-                                    /**
-                                     * S3 endpoint, e.g. `https://s3-us-west-2.amazonaws.com`. Either region or endpoint needs to be set.
-                                     * If both are set, endpoint will be used.
-                                     */
-                                    endpoint?: string;
-                                    /**
-                                     * (Optional) Kms key which will be used if encryption is enabled and encryption type is set to `sse-kms`.
-                                     */
-                                    kms_key?: string;
-                                    /**
-                                     * S3 region, e.g. `us-west-2`. Either region or endpoint needs to be set. If both are set,
-                                     * endpoint will be used.
-                                     */
-                                    region?: string;
-                                };
-                            };
-                            /**
-                             * Additional tags for cluster resources. Databricks will tag all cluster resources (e.g., AWS
-                             * instances and EBS volumes) with these tags in addition to `default_tags`. Notes:
-                             *
-                             * - Currently, Databricks allows at most 45 custom tags
-                             *
-                             * - Clusters can only reuse cloud resources if the resources' tags are a subset of the cluster tags
-                             */
-                            custom_tags?: {
-                                [k: string]: string;
-                            };
-                            /**
-                             * The optional ID of the instance pool for the driver of the cluster belongs.
-                             * The pool cluster uses the instance pool with id (instance_pool_id) if the driver pool is not
-                             * assigned.
-                             */
-                            driver_instance_pool_id?: string;
-                            /**
-                             * The node type of the Spark driver.
-                             * Note that this field is optional; if unset, the driver node type will be set as the same value
-                             * as `node_type_id` defined above.
-                             */
-                            driver_node_type_id?: string;
-                            /**
-                             * Attributes related to clusters running on Google Cloud Platform.
-                             * If not specified at cluster creation, a set of default values will be used.
-                             */
-                            gcp_attributes?: {
-                                availability?: string;
-                                /**
-                                 * boot disk size in GB
-                                 */
-                                boot_disk_size?: number;
-                                /**
-                                 * If provided, the cluster will impersonate the google service account when accessing
-                                 * gcloud services (like GCS). The google service account
-                                 * must have previously been added to the Databricks environment by an account
-                                 * administrator.
-                                 */
-                                google_service_account?: string;
-                                /**
-                                 * If provided, each node (workers and driver) in the cluster will have this number of local SSDs attached. Each local SSD is 375GB in size. Refer to [GCP documentation](https://cloud.google.com/compute/docs/disks/local-ssd#choose_number_local_ssds) for the supported number of local SSDs for each instance type.
-                                 */
-                                local_ssd_count?: number;
-                            };
-                            /**
-                             * The optional ID of the instance pool to which the cluster belongs.
-                             */
-                            instance_pool_id?: string;
-                            /**
-                             * A label for the cluster specification, either `default` to configure the default cluster, or `maintenance` to configure the maintenance cluster. This field is optional. The default value is `default`.
-                             */
-                            label?: string;
-                            /**
-                             * This field encodes, through a single value, the resources available to each of
-                             * the Spark nodes in this cluster. For example, the Spark nodes can be provisioned
-                             * and optimized for memory or compute intensive workloads. A list of available node
-                             * types can be retrieved by using the :method:clusters/listNodeTypes API call.
-                             *
-                             */
-                            node_type_id?: string;
-                            /**
-                             * Number of worker nodes that this cluster should have. A cluster has one Spark Driver
-                             * and `num_workers` Executors for a total of `num_workers` + 1 Spark nodes.
-                             *
-                             * Note: When reading the properties of a cluster, this field reflects the desired number
-                             * of workers rather than the actual current number of workers. For instance, if a cluster
-                             * is resized from 5 to 10 workers, this field will immediately be updated to reflect
-                             * the target size of 10 workers, whereas the workers listed in `spark_info` will gradually
-                             * increase from 5 to 10 as the new nodes are provisioned.
-                             */
-                            num_workers?: number;
-                            /**
-                             * The ID of the cluster policy used to create the cluster if applicable.
-                             */
-                            policy_id?: string;
-                            /**
-                             * An object containing a set of optional, user-specified Spark configuration key-value pairs.
-                             * See :method:clusters/create for more details.
-                             *
-                             */
-                            spark_conf?: {
-                                [k: string]: string;
-                            };
-                            /**
-                             * An object containing a set of optional, user-specified environment variable key-value pairs.
-                             * Please note that key-value pair of the form (X,Y) will be exported as is (i.e.,
-                             * `export X='Y'`) while launching the driver and workers.
-                             *
-                             * In order to specify an additional set of `SPARK_DAEMON_JAVA_OPTS`, we recommend appending
-                             * them to `$SPARK_DAEMON_JAVA_OPTS` as shown in the example below. This ensures that all
-                             * default databricks managed environmental variables are included as well.
-                             *
-                             * Example Spark environment variables:
-                             * `{"SPARK_WORKER_MEMORY": "28000m", "SPARK_LOCAL_DIRS": "/local_disk0"}` or
-                             * `{"SPARK_DAEMON_JAVA_OPTS": "$SPARK_DAEMON_JAVA_OPTS -Dspark.shuffle.service.enabled=true"}`
-                             */
-                            spark_env_vars?: {
-                                [k: string]: string;
-                            };
-                            /**
-                             * SSH public key contents that will be added to each Spark node in this cluster. The
-                             * corresponding private keys can be used to login with the user name `ubuntu` on port `2200`.
-                             * Up to 10 keys can be specified.
-                             */
-                            ssh_public_keys?: string[];
-                        }[];
-                        /**
-                         * String-String configuration for this pipeline execution.
-                         */
-                        configuration?: {
-                            [k: string]: string;
-                        };
-                        /**
-                         * Whether the pipeline is continuous or triggered. This replaces `trigger`.
-                         */
-                        continuous?: boolean;
-                        /**
-                         * Whether the pipeline is in Development mode. Defaults to false.
-                         */
-                        development?: boolean;
-                        /**
-                         * Pipeline product edition.
-                         */
-                        edition?: string;
-                        /**
-                         * Filters on which Pipeline packages to include in the deployed graph.
-                         */
-                        filters?: {
-                            /**
-                             * Paths to exclude.
-                             */
-                            exclude?: string[];
-                            /**
-                             * Paths to include.
-                             */
-                            include?: string[];
-                        };
-                        /**
-                         * Unique identifier for this pipeline.
-                         */
-                        id?: string;
-                        /**
-                         * Libraries or code needed by this deployment.
-                         */
-                        libraries?: {
-                            /**
-                             * The path to a file that defines a pipeline and is stored in the Databricks Repos.
-                             *
-                             */
-                            file?: {
-                                /**
-                                 * The absolute path of the file.
-                                 */
-                                path?: string;
-                            };
-                            /**
-                             * URI of the jar to be installed. Currently only DBFS is supported.
-                             *
-                             */
-                            jar?: string;
-                            /**
-                             * Specification of a maven library to be installed.
-                             *
-                             */
-                            maven?: {
-                                /**
-                                 * Gradle-style maven coordinates. For example: "org.jsoup:jsoup:1.7.2".
-                                 */
-                                coordinates: string;
-                                /**
-                                 * List of dependences to exclude. For example: `["slf4j:slf4j", "*:hadoop-client"]`.
-                                 *
-                                 * Maven dependency exclusions:
-                                 * https://maven.apache.org/guides/introduction/introduction-to-optional-and-excludes-dependencies.html.
-                                 */
-                                exclusions?: string[];
-                                /**
-                                 * Maven repo to install the Maven package from. If omitted, both Maven Central Repository
-                                 * and Spark Packages are searched.
-                                 */
-                                repo?: string;
-                            };
-                            /**
-                             * The path to a notebook that defines a pipeline and is stored in the <Databricks> workspace.
-                             *
-                             */
-                            notebook?: {
-                                /**
-                                 * The absolute path of the notebook.
-                                 */
-                                path?: string;
-                            };
-                        }[];
-                        /**
-                         * Friendly identifier for this pipeline.
-                         */
-                        name?: string;
-                        permissions?: {
-                            group_name?: string;
-                            level: string;
-                            service_principal_name?: string;
-                            user_name?: string;
-                        }[];
-                        /**
-                         * Whether Photon is enabled for this pipeline.
-                         */
-                        photon?: boolean;
-                        /**
-                         * Whether serverless compute is enabled for this pipeline.
-                         */
-                        serverless?: boolean;
-                        /**
-                         * DBFS root directory for storing checkpoints and tables.
-                         */
-                        storage?: string;
-                        /**
-                         * Target schema (database) to add tables in this pipeline to. If not specified, no data is published to the Hive metastore or Unity Catalog. To publish to Unity Catalog, also specify `catalog`.
-                         */
-                        target?: string;
-                        /**
-                         * Which pipeline trigger to use. Deprecated: Use `continuous` instead.
-                         */
-                        trigger?: {
-                            cron?: {
-                                quartz_cron_schedule?: string;
-                                timezone_id?: string;
-                            };
-                            manual?: unknown;
-                        };
-                    };
-                };
-            };
-            run_as?: {
-                service_principal_name?: string;
-                user_name?: string;
-            };
-            variables?: {
-                [k: string]: string;
-            };
-            /**
-             * Configures which workspace to connect to and locations for files, state, and similar locations within the workspace file tree.
-             */
-            workspace?: {
-                /**
-                 * The remote path to synchronize build artifacts to. This defaults to `${workspace.root}/artifacts`
-                 */
-                artifact_path?: string;
-                auth_type?: string;
-                azure_client_id?: string;
-                /**
-                 * Azure environment, one of (Public, UsGov, China, Germany).
-                 */
-                azure_environment?: string;
-                /**
-                 * Azure Login Application ID.
-                 */
-                azure_login_app_id?: string;
-                azure_tenant_id?: string;
-                azure_use_msi?: boolean;
-                /**
-                 * Azure Resource Manager ID for Azure Databricks workspace.
-                 */
-                azure_workspace_resource_id?: string;
-                client_id?: string;
-                /**
-                 * The remote path to synchronize local files artifacts to. This defaults to `${workspace.root}/files`
-                 */
-                file_path?: string;
-                google_service_account?: string;
-                /**
-                 * Host url of the workspace.
-                 */
-                host?: string;
-                /**
-                 * Connection profile to use. By default profiles are specified in ~/.databrickscfg.
-                 */
-                profile?: string;
-                /**
-                 * The base location for synchronizing files, artifacts and state. Defaults to `/Users/jane@doe.com/.bundle/${bundle.name}/${bundle.target}`
-                 */
-                root_path?: string;
-                /**
-                 * The remote path to synchronize bundle state to. This defaults to `${workspace.root}/state`
-                 */
-                state_path?: string;
-            };
-        };
-    };
-    variables?: {
-        [k: string]: {
-            default?: string;
-            description?: string;
-        };
-    };
-    /**
-     * Configures which workspace to connect to and locations for files, state, and similar locations within the workspace file tree.
-     */
-    workspace?: {
-        /**
-         * The remote path to synchronize build artifacts to. This defaults to `${workspace.root}/artifacts`
-         */
-        artifact_path?: string;
-        auth_type?: string;
-        azure_client_id?: string;
-        /**
-         * Azure environment, one of (Public, UsGov, China, Germany).
-         */
-        azure_environment?: string;
-        /**
-         * Azure Login Application ID.
-         */
-        azure_login_app_id?: string;
-        azure_tenant_id?: string;
-        azure_use_msi?: boolean;
-        /**
-         * Azure Resource Manager ID for Azure Databricks workspace.
-         */
-        azure_workspace_resource_id?: string;
-        client_id?: string;
-        /**
-         * The remote path to synchronize local files artifacts to. This defaults to `${workspace.root}/files`
-         */
-        file_path?: string;
-        google_service_account?: string;
-        /**
-         * Host url of the workspace.
-         */
-        host?: string;
-        /**
-         * Connection profile to use. By default profiles are specified in ~/.databrickscfg.
-         */
-        profile?: string;
-        /**
-         * The base location for synchronizing files, artifacts and state. Defaults to `/Users/jane@doe.com/.bundle/${bundle.name}/${bundle.target}`
-         */
-        root_path?: string;
-        /**
-         * The remote path to synchronize bundle state to. This defaults to `${workspace.root}/state`
-         */
-        state_path?: string;
-    };
-}
