--- packages/databricks-vscode/src/bundle/types.ts
@@ -1,3 +1,0 @@
-import {BundleSchema} from "./BundleSchema";
-
-export type BundleTarget = Required<BundleSchema>["targets"][string];
