+++ packages/databricks-vscode/src/cli/DatabricksCliSyncParser.ts
@@ -99,7 +99,6 @@ export class DatabricksCliSyncParser {
                 return;
             }
         }
-        this.syncStateCallback("ERROR", data);
     }
 
     private matchForErrors(line: string) {
