+++ packages/databricks-vscode/src/configuration/auth/DatabricksCliCheck.ts
@@ -2,12 +2,10 @@ import {
     ExecUtils,
     ProductVersion,
     WorkspaceClient,
-    logging,
 } from "@databricks/databricks-sdk";
 import {Disposable, window} from "vscode";
 import {DatabricksCliAuthProvider} from "./AuthProvider";
 import {orchestrate, OrchestrationLoopError, Step} from "./orchestrate";
-import {Loggers} from "../../logger";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const extensionVersion = require("../../../package.json")
@@ -25,7 +23,7 @@ export class DatabricksCliCheck implements Disposable {
         this.disposables = [];
     }
 
-    async check(): Promise<boolean> {
+    async check(silent: boolean): Promise<boolean> {
         const steps: Record<StepName, Step<boolean, StepName>> = {
             tryLogin: async () => {
                 if (await this.tryLogin()) {
@@ -57,15 +55,12 @@ export class DatabricksCliCheck implements Disposable {
             } else {
                 message = e.message;
             }
-            logging.NamedLogger.getOrCreate(Loggers.Extension).error(
-                message,
-                e
-            );
+
             window.showErrorMessage(message);
             return false;
         }
 
-        if (result) {
+        if (result && !silent) {
             window.showInformationMessage(
                 "Databricks: Successfully logged in with Databricks CLI"
             );
