+++ packages/databricks-vscode/src/configuration/auth/orchestrate.ts
@@ -42,7 +42,7 @@ export async function orchestrate<S, KEYS extends string>(
             throw new OrchestrationLoopError();
         }
         const result: StepResult<S, KEYS> = await steps[step]();
-        logger?.info(`Auth check: ${step}`, result);
+        logger?.info(`Azire CLI check: ${step}`, result);
 
         if (result.type === "error") {
             throw result.error;
