+++ packages/databricks-vscode/src/locking/index.ts
@@ -1,9 +1,2 @@
-import {Mutex} from "./Mutex";
-
 export * from "./Mutex";
 export * from "./RWLock";
-
-export class WithMutex<T> {
-    public readonly mutex = new Mutex();
-    constructor(public value: T) {}
-}
